Function of collecting articles on many press of Burkina Faso

Collection function of all the articles of contry Burkina Faso

bf_news_collect(start_d, end_d, press = "all")

Arguments

start_d
yyyy-mm-dd start date of collection (oldest date, eg: yesterday)

end_d
yyyy-mm-dd end date of collection (newest date, eg: today)

press
specify the name of a press.

Value

a table containing the articles published in the period

Examples

library(afnews)
dt_ <- bf_news_collect(start_d="2020-10-20", end_d="2020-10-28", press="all")

#> Done!
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> Done!

head(dt_)

#>   date_article
#> 1   2020-10-22
#> 2   2020-10-21
#> 3   2020-10-20
#> 4   2020-10-28
#> 5   2020-10-28
#> 6   2020-10-27
#>                                                                                                                     url_article
#> 1                                                                                      http://lefaso.net/spip.php?article100134
#> 2                                                                                      http://lefaso.net/spip.php?article100129
#> 3                                                                                      http://lefaso.net/spip.php?article100101
#> 4 https://www.burkina24.com/2020/10/28/cote-divoire-tidjane-thiam-rejoint-lopposition-contre-la-candidature-dalassane-ouattara/
#> 5                      https://www.burkina24.com/2020/10/28/cote-divoire-la-presidentielle-aura-bel-et-bien-lieu-le-31-octobre/
#> 6            https://www.burkina24.com/2020/10/27/guinee-une-mission-conjointe-onu-ua-cedeao-a-rencontre-cellou-dallein-diallo/
#>                                                                                                          title_article
#> 1 Appel de Gaoua sur la qualité de l’enseignement au Burkina Faso : Trente-quatre ans après, des réalités similaires ?
#> 2                          Culture et démocratie : Comment concilier les deux dans le cadre de la cohésion nationale ?
#> 3                                                        Financement du cinéma burkinabè : La Grande marche à Reculons
#> 4                         Côte d’Ivoire : Tidjane Thiam rejoint l’opposition contre la candidature d’Alassane Ouattara
#> 5                                              Côte d’Ivoire : «La présidentielle aura bel et bien lieu le 31 octobre»
#> 6                                       Guinée : Une mission conjointe ONU-UA-CEDEAO a rencontré Cellou Dallein Diallo