Main functions of collecting news from Burkina Faso

Function for collecting news (title + content) from the media in Burkina Faso.

bf_news_contents(start_d, end_d, press = "all")

Arguments

start_d
yyyy-mm-dd start date of collection (oldest date, eg: yesterday)

end_d
yyyy-mm-dd collection end date (newest date, eg: today)

press
specify the name of a press.

Value

a table containing the news (title + content + date publisher + press) collected

Examples

library(afnews)
dt_news <- bf_news_contents(start_d="2020-10-20", end_d="2020-10-28", press='all')

#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> Done!
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> Done!

head(dt_news)

#>   date_article                              url_article
#> 1   2020-10-25 http://lefaso.net/spip.php?article100202
#> 2   2020-10-23 http://lefaso.net/spip.php?article100172
#> 3   2020-10-22 http://lefaso.net/spip.php?article100132
#> 4   2020-10-22 http://lefaso.net/spip.php?article100145
#> 5   2020-10-21 http://lefaso.net/spip.php?article100123
#> 6   2020-10-21 http://lefaso.net/spip.php?article100114
#>                                                                                                               title_article
#> 1                                            Dori : Le Premier ministre échange avec les forces vives de la région du Sahel
#> 2        Liberté provisoire de Jean Claude Bouda : "Cela nous intrigue...", Sagado Nacanabo, Secrétaire exécutif du REN-LAC
#> 3                            Burkina : Pascal Zaida et Marcel Tankoano alertent sur les risques d’une crise post-électorale
#> 4                               Justice : Jean Claude Bouda en liberté provisoire, 5 mois après son incarcération à la MACO
#> 5       Réconciliation nationale : « Pourquoi avoir une haine au point de ne pas pardonner ? », Dr Justine Couldiati/Kielem
#> 6 Roch Kaboré et la Réconciliation nationale : Le Mouvement pour le retour de Blaise Compaoré y voit un affront, un mépris…
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        Content_article
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  C’est aux environs de 10h30 que le chef du gouvernement burkinabè a foulé le sol de la région du Sahel. Après un accueil chaleureux par les autorités de la région à l’entrée de Dori, Christophe Dabiré et sa délégation ont été conduits à la salle polyvalente de Dori pour ce face-à-face avec les forces vives de la région.\nAprès le mot de bienvenue du député-maire de Dori, Aziz Diallo, le gouverneur de la région du Sahel, le colonel-major Salfo Kaboré, le président du Conseil régional du Sahel, le président du Conseil régional des jeunes ainsi que la coordinatrice des femmes du Sahel et d’autres intervenants, ont égrainé les problèmes qui empêchent le développement de la région du Sahel. De l’insécurité en passant par la fermeture de certaines écoles et de centres de santé, rien n’a été oublié par les intervenants.\n\nLe face à face entre le premier ministre et les forces vives du sahel \nToutefois, ils n’ont pas manqué de saluer les efforts du gouvernement à travers les différentes réalisations, malgré l’insécurité qui a empêché l’avancement des travaux de développement dans la région du Sahel.\nEn réponse, Christophe Dabiré a rassuré que le gouvernement fera tout pour la sécurisation de la région, afin de permettre la reprise des réalisations qui sont bloquées à cause de l’insécurité. Le Premier ministre a ajouté que les préoccupations citées par la population qui n’ont pas été prises en compte par l’ancien référentiel, le seront dans le nouveau référentiel de développement en cours d’élaboration. Christophe Dabiré a aussi rappelé aux populations qu’elles ne sont pas oubliées par le gouvernement ; d’où sa présence pour leur porter le message du président du Faso qui est celui de l’espoir, du vivre-ensemble dans la cohésion sociale.\n\nEchantillon des 80 tonnes de vivre que le premier ministre a offert au deplacés internes de Dori \nLa région du Sahel enregistre de nombreux déplacés internes. Ces derniers ont reçu 80 tonnes de vivres de la part du Premier ministre.\nIl faut noter que ce face-à-face avec les forces vives de la région du Sahel s’inscrit dans une série de prises de contact avec toutes les forces vives des treize régions du Burkina Faso. Ces rencontres, a dit Christophe Dabiré, visent à recueillir les préoccupations des populations pour la rédaction du nouveau révérenciel de développement du Burkina Faso.\nPinidié Gnanou \nLefaso.net
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          Lefaso.net : Comment le REN-LAC a accueilli la nouvelle de la liberté accordée à Jean Claude Bouda ?\nSagado Nacanabo : L’affaire BOUDA est un dossier qui nous tient beaucoup à cœur au niveau du REN-LAC. Elle fait partie des dossiers qu’on appelle couramment les dossiers emblématiques, au regard de la personnalité de monsieur BOUDA. Sa libération, nous l’avons apprise comme tout le monde à travers la presse et les réseaux sociaux. On peut dire qu’on est un peu surpris, mais on comprend qu’en droit, en attendant le procès, la règle c’est la liberté, et l’exception, c’est la détention.\nLe 19 juillet 2020 dans un communiqué signé du secrétaire exécutif, le REN-LAC réclamait le jugement du dossier au fond. Avec cette décision de la Cour d’Appel, le réseau espère-t-il toujours de voir le dossier jugé tel qu’il le souhaite ?\nBien sûr ! cette liberté à lui accordée ne met pas fin à la procédure judiciaire. Généralement le prévenu est placé en détention provisoire pour des raisons d’enquête. Si le juge estime qu’il n’y a plus de raisons de le maintenir en prison, il peut le libérer. Mais pour le REN-LAC, comme on l’a dit dans le communiqué du 19 juillet, l’essentiel pour nous c’est que ce dossier soit jugé au fond, et nous ne désespérons pas. D’autant plus que ce matin même du vendredi 23 octobre, nous avons été auditionné par le juge d’instruction dans le cadre de cette affaire.\nAvez-vous déjà envisagé que, si procès il y a sur le fond du dossier, la justice prononce sa relaxe ? Si oui ou non, pourquoi ?\nComme dans tout procès, il y a essentiellement deux hypothèses, soit vous êtes reconnu coupable et vous écopez d’une peine, soit vous n’êtes pas reconnu coupable et on vous relaxe. Mais, pour nous, à la lumière des éléments du dossier, et en tant que structure ayant porté plainte et également partie civile, nous pensons qu’il sera condamné pour les faits à lui reprochés afin de donner un signal fort pour la lutte contre la corruption.  De toutes les façons, à l’issue du procès si une des parties n’est pas satisfaite, elle peut faire appel. Et même un pourvoi en cassation si cela est nécessaire.\nLire aussi : Justice : Jean Claude Bouda en liberté provisoire, 5 mois après son incarcération à la MACO\nComment analysez-vous le fait que le 29 septembre 2020, Jean Claude Bouda ait bénéficié d’une permission pour des motifs de santé et quelques jours après, la Cour prononce une liberté en sa faveur ?\nCela nous intrigue que ce soit pendant sa période de permission pour des soins qu’il bénéficie d’une liberté ordonnée par la chambre d’instruction en appel. Nous ne faisons pas une fixation particulière sur la liberté de M. Bouda. Nous restons plutôt sur notre position qui est que l’instruction en cours du dossier puisse aboutir à un jugement au fond.\nLa décision tombe à quelques jours des élections couplées du 22 novembre prochain. Faut-il y voir une main politique dernière ? Dans lequel cas, cela remet en doute l’indépendance de la Justice.\nDe manière générale, on ne peut pas dire que notre justice est totalement indépendante. Il est aussi évident que cette décision à quelques jours des élections peut avoir des retombées politiques, certes, mais le plus important est que le procès ait lieu. Et nous espérons que cette affaire sera une occasion pour la justice burkinabè d’affirmer davantage son indépendance.\nLefaso.net
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  A une semaine de la campagne électorale et juste 24h après les résultats de l’audit du fichier électoral, la société civile monte au créneau pour alerter sur les risques d’une crise post-électorale. Pascal Zaida du cadre d’expression démocratique et Marcel Tankoano du M21 déplorent une non couverture géographique du territoire national avec 22 communes exclues de l’enrôlement et la « non prise de sanctions contre les tentatives de fraude lors de l’enrôlement ».\nSelon Pascal Zaida, dans un village de la commune de Komsilga, composé de 300 habitants, plus de 800 personnes ont été enrôlées. Et à l’en croire, ce n’est pas demain la veille qu’aboutira le processus d’audition engagé par les forces de sécurité. « Il faut que les populations fassent barrage à tout cela. Que personne ne quitte ailleurs pour venir voter dans leur localité (...) Il faut que la classe politique accepte dire aux gens qui se sont fait enrôler ailleurs qu’ils ne pourront pas aller voter dans ces zones », conseille Pascal Zaida.\n\nPascal Zaida du Cadre d’expression démocratique \n« Les risques »\nPour les conférenciers, plusieurs risques pèsent sur le processus électoral. Il s’agit selon eux de l’existence des bureaux de vote fictifs, des urnes fictives, et des bulletins pré paraphés. Autre risque, selon le CED et le M21, ce sont les velléités d’utilisation de téléphones portables pour que les électeurs rapportent la preuve de leur vote dans l’isoloir. « La récupération des cartes d’électeurs par des tierces personnes et le convoi des électeurs vers des lieux de vote » ont été également dénoncés par les conférenciers. Et avant de lancer la campagne, le CED et le M21 estiment que ces prémices peuvent être l’objet de contestations des résultats.\nAppel au sens de la responsabilité et à la probité\n« C’est toute la classe politique qui sera responsable du chaos que le Burkina va vivre », a mis en garde Pascal Zaida. Il a joint sa voix à celle de Marcel Tankoano pour appeler les différents acteurs « au sens de la responsabilité et de la probité ». Aux institutions comme la Commission électorale nationale indépendante (CENI) et le Conseil constitutionnel, les conférenciers demandent « la vigilance et l’impartialité ». Les partis politiques et les candidats ont été invités à être « rigoureux » dans le choix de leurs représentants dans les bureaux de vote.\n\nMarcel Tankoano du Mouvement M21 \n« On ne soutient personne mais nous allons demander aux gens d’aller faire un vote utile », a clarifié Pascal Zaida en disant que le moment voulu les deux OSC donneront des orientations aux citoyens pour le vote de celui qui répond le plus aux aspirations de la majorité.\nHerman Frédéric Bassolé\nLefaso.net
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     Jean Claude Bouda, ex ministre de la défense, était mis en examen pour  faux et usage de faux en écriture privée, délit d’apparence...\nLa Cour a ordonné la mise en liberté de l’ex ministre de la défense "s’il n’est retenu pour autre chose".\nIl y a quelques jours, soit le 29 septembre 2020, Jean Claude Bouda avait bénéficié d’une "permission" d’un mois pour des motifs de santé. Cette permission a dû faciliter la mise en liberté provisoire dont il bénéficie aux terme d’une décision de la Chambre de l’instruction de la Cour d’appel en date de ce 22 octobre.\nOn se rappelle aussi que l’avocat de M. Bouda a saisi la Cour de cassation sur  la compétence du tribunal connaissant de l’affaire. En attendant la décision de la Cour de cassation, Jean-Claude Bouda, qui bénéficiait déjà d’une permission de sortie pour des soins, est donc en liberté provisoire à compter de ce 22 octobre.\nLefaso.net\nPour tout savoir de l’affaire Jean Claude Bouda, la rédaction vous propose : Jean-Claude Bouda déposé à la Maison d’arrêt et de correction de Ouagadougou\nAffaire Jean Claude Bouda : L’ancien ministre de la défense « en permission » d’un mois pour des soins \nLire aussi Affaire Jean-Claude Bouda : Le RENLAC réclame le jugement du dossier au fond
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 De sa motivation pour la présidence du Faso, son regard sur la situation de crise au Burkina et dans l’espace sahélo-saharien, le drame de Yirgou, le besoin de la réconciliation nationale..., la sociologue ne voile pas ses mots pour dénoncer, condamner, apprécier ou proposer.\n« L’exclusion n’a jamais aidé une famille nucléaire, à plus forte raison un pays. Même si vous réunissez 90% de la population autour de la table familiale, les 5,10% que vous allez exclure, même s’ils sont conscients que les décisions prises sont acceptables, mais le fait qu’ils soient exclus, ils n’accepteront jamais. Que nous nous retrouvions autour d’une table, les uns et les autres reconnaissent ce qu’ils ont fait de mauvais, mais qu’on leur reconnaisse aussi ce qu’ils ont fait de bien.\nJe vous prends un exemple : la guerre de Noël, où était notre papa Saye Zerbo (chef de l’Etat de la Haute-Volta du 25 novembre 1980 au 7 novembre 1982, ndlr) ? En prison. Lorsqu’on est parti le chercher, on l’a amené là où il fallait faire les plans de guerre et il a travaillé avec tout le monde et lorsque ça a marché, qu’est-ce qu’on a fait ? On l’a ramené chez lui, la prison était finie. Pourquoi avoir une haine au point de ne pas pardonner ? Est-on chrétien ? Est-on musulman ?", s’est appesantie Dr Justine Couldiati/Kielem.\nAncienne présidente de la plateforme des femmes du G5 Sahel, elle pense que ce n’est pas par les bureaux que l’organisation sous-régionale de lutte contre le terrorisme parviendra à son objectif. « Je vais demander à tous ceux qui contribuent dans le G5 Sahel, combien vont dans l’armement, combien vont dans la formation des forces spéciales et combien vont dans les missions. C’est triste. On ne peut pas demander à des pays pauvres (Mali, Niger, Burkina, Mauritanie, Tchad) de contribuer à hauteur de deux milliards par an, et plus d’un milliard cinq cent millions, c’est pour s’occuper de ceux qui sont dans les bureaux ; leurs émoluments et leurs enfants. Que voulez-vous qu’on fasse ? Et dire aux gens d’amener l’argent pour gérer les questions sécuritaires, quand on sait que certains mêmes (partenaires étrangers, ndlr) sont peut-être complices de la situation que nous vivons. Je ne pense pas que le G5 Sahel peut gérer quoi que ce soit », s’insurge Dr Justine Couldiati. Suivez sur Lefaso TV, l’intégralité de cette interview qu’elle nous a accordée le 10 octobre 2020.\n\nEdouard Samboé \nOumar Ouédraogo\nAdjaratou Tall\nLefaso.net
#> 6 En Afrique, il y a un adage qui dit : « mieux vaut tard que jamais ». Mais il était essentiel pour nous, de faire la genèse de cette question de réconciliation et du retour des exilés qui se retrouve aujourd’hui dans des débats politiques à quelques jours des élections présidentielle et législatives.\nDepuis 2014, soit plus de 6 ans, les clivages politiques exacerbés par une chasse aux sorcières règnent en maîtres au Burkina Faso.\nNous avons espéré que le régime de Roch Marc Christian KABORE se revêtirait de la tunique de sagesse des autorités Sud-africaines qui, malgré les atrocités et les barbaries de l’apartheid, ne se sont pas abreuvées dans la source destructrice de la haine, de la rancune et du règlement de compte tous azimuts. Mais hélas ! Les autorités aux commandes n’en ont eu cure de la nécessité de la réconciliation pour impulser un véritable développement économique et social dont le seul secret est l’unité nationale. Pourtant,  des voix n’ont pas cessé d’appeler le président à la sagesse pour œuvrer à la réconciliation et au retour des Exilés.\nMais que c’est triste, rien ne fit ! Tout au contraire, le président du Faso, à travers plusieurs sorties médiatiques,  n’a pas cessé d’associer le nom de ses adversaires politiques aux œuvres barbares des terroristes qui profitent des clivages politiques et sociaux pour instaurer un lit de désolation aux préjudices de nos braves et vaillants concitoyens des villes et campagnes.\nEn dépit des railleries, des supercheries et de l’indifférence avilissante du régime de Roch KABORE, nous avons toujours appelé de tous nos vœux, à des actions concrètes du gouvernement concernant le retour des exilés, au traitement décent des prisonniers politiques et à la réconciliation, ce, passant par nos actions de conférences de presse, d’interpellations de toute la Nation en général et du Président du Faso en particulier. Alors sur les visites des chefs coutumiers et religieux, nous n’en comptons pas le nombre de plaidoyers de 2015 à nos jours.\nNotons encore que sur toutes nos sorties médiatiques, plateaux télé comme conférences de presse, conférences publiques et conférences débats, nous avons toujours interpelé le président Roch KABORE sur la nécessité de l’unité nationale pour un Burkina vrai, un Burkina des intègres.\nIl sera intéressant de noter qu’il y a plus de 50 structures OSC qui travaillent depuis toujours sur la question de réconciliation nationale et du retour des exilés.\nSelon l’anthropologue Georges Samuelson, lorsqu’un groupe divisé est sous l’emprise d’une menace exogène, il s’unit instinctivement pour se défendre contre ledit facteur extérieur constituant une menace à sa survie. Cet instinct habite nombre de nos concitoyens qui de manière organisée ou pas appellent à l’unité.\nEn effet, que par voie de presse, et à l’unisson avec des organisations telles le CED de Pascal ZAIDA, le CDR de NANA Thibaut, la CODER, le MPRCN de Safiatou Lopez, l’opposition politique Burkinabè par le CFOP et l’Appel de Manéga qui a d’ailleurs proposé une feuille de route sur le processus de réconciliation pour qu’enfin on parvienne à une véritable réconciliation Nationale pour l’intérêt de tous les Burkinabè.\nToutes les propositions émanant de ces structures ont été balayées du revers de la main par le régime KABORE qui n’a eu de cesse de fredonner leur refrain préféré : « Les frontières ne sont pas fermées ».\nAlors, en somme, nous avons compris qu’il est difficile de réveiller quelqu’un qui ne dort pas.\nUn régime qui était, hostile, farouche à toute idée de réconciliation et surtout des exilés, se livre à une volte-face depuis peu, à quelques jours des échéances électorales. Le régime vient d’inscrire désormais le  thème de réconciliation nationale dans son agenda pour 2021 et ce, conditionné par la réélection de Roch Marc Christian KABORE. Cela est un affront, un mépris et une insulte à la conscience collective nationale et à la mémoire de tout le peuple. Durant 5 années, le régime s’est rendu sourd à la question de réconciliation et du retour des exilés qui est sans doute notre champ de bataille pour un Burkina uni.\nMaintenant que l’heure des élections a sonné, prêt à tout compromis pour se maintenir au pouvoir, le régime, honteusement et par un simulacre de bonne volonté  et un jeu théâtral,  appelle à la réconciliation nationale.\nLe président Roch Marc Christian KABORE, toute la nation épouse votre sens élevé de responsabilité pris sur le processus de réconciliation et du retour des exilés lors de votre interview accordée à France24 et RFI le 15 Octobre 2020. Mais conditionner cela à votre réélection est une insulte collective portée à toute la Nation burkinabè.\nNous restons confiants, déterminés, vigilants et soutenons tout engagement et propos de réconciliation nationale et du retour des exilés.\nVive la réconciliation des filles et fils du Burkina, \nVive les concitoyens des villes et campagnes en lutte contre l’hydre du terrorisme, \nVive le peuple déterminé à défendre sans concession l’intérêt de la Nation,\nVive les organisations en tandem luttant pour la Cohésion et le Développement.\n Je vous remercie !\nOuagadougou, le 20/10/2020\nLe Coordinateur du Mouvement Pour le Retour du Président Blaise Compaoré et des Exilés.\nMamadou Abdel Kader TRAORE\nLire aussi :Réconciliation nationale : « Nous allons régler la question dès le premier semestre 2021 », promet Roch Kaboré sur RFI
#>   press_article
#> 1       fasonet
#> 2       fasonet
#> 3       fasonet
#> 4       fasonet
#> 5       fasonet
#> 6       fasonet