Main functions of collecting information on Benin

Function for collecting news (title + content) from the media in Benin.

bn_news_contents(start_d, end_d, press = "all")

Arguments

start_d
yyyy-mm-dd start date of collection (oldest date, eg: yesterday)

end_d
yyyy-mm-dd collection end date (newest date, eg: today)

press
specify the name of a press.

Value

a table containing the news (title + content + date publisher + press) collected

Examples

library(afnews)
dt_news <- bn_news_contents(start_d="2020-11-01", end_d="2020-11-10", press='all')

#> Done!

#> Warning: No article publish in this period !

#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> Done!

head(dt_news)

#>   date_article
#> 1   2020-11-10
#> 2   2020-11-10
#> 3   2020-11-10
#> 4   2020-11-09
#> 5   2020-11-09
#> 6   2020-11-09
#>                                                                                                                                                  url_article
#> 1                                                                           https://lanationbenin.info/match-benin-lesotho-soukou-adeoti-et-verdon-forfaits/
#> 2                                                     https://lanationbenin.info/utilisation-de-la-plateforme-sigmap-les-acteurs-des-marches-publics-formes/
#> 3                                    https://lanationbenin.info/projet-de-budget-de-letat-gestion-2021-romuald-wadagni-presente-les-grands-axes-aux-deputes/
#> 4                                               https://lanationbenin.info/felicitations-a-biden-ces-chefs-detat-qui-nattendent-que-les-resultats-officiels/
#> 5 https://lanationbenin.info/marcellin-laourou-inspecteur-de-douane-toute-marchandise-saisie-non-retiree-dans-un-delai-de-15-jours-sera-vendue-aux-encheres/
#> 6                                         https://lanationbenin.info/preparatifs-du-match-benin-lesotho-14-joueurs-expatries-au-bercail-six-attendus-demain/
#>                                                                                                                           title_article
#> 1                                                                               Match Bénin-Lesotho : Soukou, Adéoti et Verdon forfaits
#> 2                                                           Utilisation de la plateforme Sigmap: Les acteurs des marchés publics formés
#> 3                                        Projet de budget de l’Etat gestion 2021 : Romuald Wadagni présente les grands axes aux députés
#> 4                                                  Félicitations à Biden : Ces chefs d’Etat qui n’attendent que les résultats officiels
#> 5 Marcellin Laourou, inspecteur de douane : « Toute marchandise saisie non retirée dans un délai de 15 jours sera vendue aux enchères »
#> 6                                              Préparatifs du match Bénin-Lesotho: 14 joueurs expatriés au bercail, six attendus demain
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               Content_article
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        Le Bénin sera privé de trois de ses cadres lors de son prochain match contre le Lesotho, samedi 14 novembre prochain à Porto-Novo.  Il s’agit de  Cèbio Soukou, Jordan Adeoti et Olivier Verdon retenus par leurs clubs pour différentes raisons liées à la pandémie du coronavirus. En effet, le club de Cèbio Soukou a refusé de le libérer car une fois de retour en Allemagne après la période FIFA du 9 au 17 novembre 2020, le joueur devra observer automatiquement la quarantaine. Cette situation va pénaliser le club en championnat car il manquera la journée du championnat juste après la trêve internationale. Idem pour Jordan Adeoti dont les responsables en clubs sont restés dans la même logique. En ce qui concerne Olivier Verdon, il s’est blessé jeudi avant d’être testé positif, ce lundi dernier, au coronavirus. Signalons que les Écureuils font leur deuxième séance d’entraînement, ce mardi à 17 h (GMT+1) au stade Charles de Gaulle à Porto-novo.
#> 2 Plus de 400 acteurs de la chaine des marchés publics sont formés, depuis lundi 9 novembre, à l’utilisation de la plateforme Sigmap à Cotonou. Organisées par l’Agence des Services et des Systèmes d’Information (Assi) et la direction nationale du contrôle des marchés publics, ces 14 sessions de formation permettront d’outiller le plus grand nombre d’utilisateurs en vue de la vulgarisation du logiciel dans l’administration publique. Quatorze sessions de formation d’une durée de cinq jours sont organisées depuis lundi 9 novembre à Cotonou, au profit des acteurs de la chaine des marchés publics en vue de l’utilisation optimale de la plateforme Sigmap.\nThierry Ahouassou, directeur des projets, infrastructures et systèmes de l’Assi, s’est réjoui de la tenue de cette formation qui favorisera l’utilisation progressive du logiciel Sigmap dans les administrations des ministères, des préfectures et celles des communes mais aussi renforcera l’accès du public à l’information relative aux marchés publics. Selon lui, cette plateforme offre des chances équitables aux soumissionnaires des marchés publics. « Il s’agit d’une application qui permet aux soumissionnaires d’avoir équitablement accès à l’information relative aux marchés publics », a-t-il déclaré. A l’en croire, il est important de valoriser le système des marchés publics qui a une incidence significative sur l’efficience dans l’utilisation des fonds publics.\nAbondant dans le même sens, Maman Adjélé Aminou, directeur général du Contrôle des marchés publics, salue cette initiative qui vise à outiller le plus grand nombre d’utilisateurs des parties contractantes afin de vulgariser la plateforme Sigmap. Selon lui, ceci témoigne de la volonté du gouvernement d’apporter des changements dans le processus de passation des marchés publics. Il a exprimé sa gratitude au gouvernement allemand qui a financé une série d’activités organisées par la direction nationale de passation des marchés publics et l’Assi. « Nous avons procédé à l’organisation d’une série d’activités dont le déploiement de la nouvelle version de la plateforme Sigmap », a-t-il déclaré.\nPour Joël Zodjhoué, secrétaire général du ministère de l’Economie et des Finances, le gouvernement oeuvre, depuis plusieurs années, à doter la chaîne des marchés publics d’outils modernes de traitement des dossiers, sous le leadership du président Patrice Talon. Dans cette optique, un appui à la maintenance évolutive du logiciel, des formations sur le Sigmap pour les utilisateurs ainsi que l’équipement des nouvelles directions départementales, notamment en matériel informatique ont été réalisés. Invitant les participants à faire preuve de discipline pour tirer meilleur profit de cette formation, il a rappelé le vœu du gouvernement d’améliorer l’efficacité et la transparence de l’administration publique dans la gestion des finances publiques.
#> 3                                                                                          Le ministre de l’Economie et des Finances, Romuald Wadagni a présenté ce lundi, aux députés membres de la Commission budgétaire, les grands axes du projet de budget de l’Etat pour la gestion 2021. La commission budgétaire de l’Assemblée nationale est désormais à fond dans l’étude du projet de loi de finances gestion 2021. Les grands axes du document ont été présentés, ce lundi 9 novembre, par le ministre de l’Economie et des Finances, Romuald Wadagni. L’argentier national est revenu sur le contenu du projet de loi de finances qui s’équilibre en ressources et en charges à la somme de 2452,192 milliards F Cfa contre 2167,099 milliards F Cfa pour la loi de finances rectificative 2020. Il annonce l’amorce du retour à la discipline budgétaire après la dégradation induite par la crise du Covid-19. La présentation du ministre Romuald Wadagni renseigne que de nombreuses mesures fiscales notamment de soutien aux entreprises, aux ménages et d’allègement du système fiscal sont contenues dans le projet de loi de finances. Il s’agit entre autres de la réforme de recouvrement de la Taxe de développement local (Tdl) sur les produits miniers ; l’exonération de la patente et de la Taxe professionnelle synthétique (Tps) sur les douze premiers mois d’activités des entreprises nouvelles ; l’exonération des rémunérations de stage de l’impôt sur le revenu des personnes physiques et du versement patronal sur salaire (Vps) ; l’assouplissement de l’amende appliquée en cas de manquements aux obligations en matière de retenue de l’Aib ; la suppression de l’obligation de demandes préalables de validation des Aib et la suspension de la retenue à la source de la Tva pour certaines entreprises en situation créditrice justifiée. Par ailleurs, le projet de budget gestion 2021 a institué le paiement fractionnel de la patente complémentaire. Cette taxe communale est exigée en totalité des entrepreneurs dès la signature du contrat et même si l’Etat n’a pas payé d’avance.\nOutre l’allègement fiscal, le projet de loi de finances 2021 ambitionne de consolider les investissements avec une politique budgétaire plus lisible et orientée vers l’investissement et le social. Il prend en compte dans ce cadre tous les secteurs d’activités, notamment l’éducation avec la poursuite de la mesure de gratuité de l’enseignement maternel et primaire ainsi que dans le secondaire pour les filles, les cantines scolaires et allocations et aides universitaires, la santé et la promotion de l’emploi des jeunes. Une priorité sera également accordée en 2021 à la dématérialisation des procédures, à l’appréhension des niches fiscales, la simplification des procédures et l’amélioration de la notation de Doing Business, soutient le ministre Romuald Wadagni, entouré des cadres techniques de son ministère.
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    A l’opposé des chefs d’Etat qui ont choisi très tôt d’adresser leurs félicitations au futur président des Usa, d’autres comme le Russe et le Chinois, préfèrent attendre les résultats officiels de l’élection avant de congratuler le vainqueur. Selon le Kremlin, cité par plusieurs médias internationaux, Vladimir Poutine attend l’annonce officielle des résultats de la présidentielle américaine avant de féliciter le vainqueur. « Nous estimons qu’il est correct d’attendre les résultats officiels des élections qui ont eu lieu. Je veux rappeler que le président Poutine a plusieurs fois dit qu’il respectera le choix du peuple américain quel qu’il soit », a déclaré le porte-parole de la présidence russe, Dmitri Peskov, aux hommes des médias. A la question de savoir pourquoi le président russe avait félicité Donald Trump un peu plus d’une heure après l’annonce de sa victoire en 2016, le porte-parole du Kremlin a répondu qu’ « à l’époque, il n’y avait pas eu de contestation juridique » des résultats. Comme la Russie, la Chine veut aussi attendre que les choses soient officielles et définitivement tranchées. « Nous avons pris note que M. Joe Biden a déclaré être le vainqueur de l’élection », affirme un porte-parole du ministère chinois des Affaires étrangères, Wang Wenbin. « A notre connaissance, ajoute-t-il, les résultats de l’élection seront déterminés par les lois et les procédures en vigueur aux Etats-Unis ». Attendre les résultats officiels, c’est également la position du Mexique. « Nous allons attendre que toutes les questions légales soient résolues. Nous ne voulons pas être imprudent », a déclaré le chef de l’Etat, Andrés Lopez Obrador. « Nous ne souhaitons pas agir à la légère et voulons être respectueux de l’autodétermination des peuples et des droits d’autrui… », a ajouté le président mexicain. Pour la Turquie, que Donald Trump ou Joe Biden soit président des Etats-Unis, cela ne changera rien entre leurs deux pays. « Du point de vue de la Turquie, est-ce que quelque chose changera ? Non, cela ne changera pas », a déclaré le vice-président turc Fuat Oktay, dans une interview à la chaîne Kanal 7. Contrairement à ces pays qui disent attendre les chiffres définitifs, le Brésil est jusque-là resté muet sur le sujet.
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     Le chef brigade du Service d’intervention rapide de la Douane et Droits indirects, Marcellin Laourou, annonce la vente aux enchères publiques des marchandises saisies par les services de douane et non encore retirées conformément à la réglementation. Après quinze jours, ces produits seront vendus aux enchères publiques, informe l’inspecteur de douane, Marcellin Laourou. « En matière douanière, les marchandises que nous saisissons, si les propriétaires ne viennent pas les retirer dans un délai de quinze jours, c’est la confiscation d’office conformément à l’article 280 du Code douanier », a indiqué l’inspecteur de douane, chef brigade du Service d’intervention rapide Atlantique-Littoral de la Douane, Marcellin Laourou. Au terme du quinzième jour, il y a confiscation d’office en dépôt de marchandises saisies par la Douane, détaille l’inspecteur de douane, pour demander à tous les usagers dont les biens se retrouvent entre les mains des disciples de Saint Mathieu à faire les formalités de retrait. Toute chose pouvant leur éviter des surprises désagréables. « Après quinze jours, la marchandise ne vous appartient plus, l’État va faire la procédure pour la vente aux enchères publiques », ajoute-t-il. Cette mise en garde concerne les propriétaires de véhicules saisis par les services de la douane. « Ceux qui ont des véhicules au niveau du Service d’intervention rapide, Atlantique-Littoral, ceux qui ont des motos et toutes sortes de marchandises que nous avons saisies, s’ils ne viennent pas les retirer dans un délai de quinze jours, nous allons les vendre aux enchères publiques, nous ne sommes pas un dépôt ordinaire pour garder des objets saisis au-delà de quinze jours », informe le douanier au cours d’un échange avec la presse à Agon, dans la commune de Toffo.
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                Les Écureuils effectueront  ce soir à Cotonou (17 h) leur première séance d’entrainement dans le cadre des préparatifs du match Bénin-Lesotho. En prélude à cette rencontre, 15 joueurs sont en regroupement depuis dimanche soir dans un hôtel à Cotonou. Il s’agit de Jodel Dossou, Cédric Hountondji, Sessi D’almeida, Johan Roche, Emmanuel Imorou, Youssouf Assogba, Chabel Gomez, Seydou Barazé,  Rodrigue Kossi, Séibou Mama, David Djigla, Khaled Adenon arrivés ce dimanche aux environs de 21 heures par vol Air France ainsi que Michael Poté venu quelques heures plus tard.  A ce groupe, s’ajoute Marcel Dandjinou, seul joueur local dans l’effectif.  Jérôme Bonou est quant à lui, attendu à Cotonou ce jour. Le reste de l’effectif, six joueurs devront rejoindre le groupe demain mardi.
#>   press_article
#> 1 lanationbenin
#> 2 lanationbenin
#> 3 lanationbenin
#> 4 lanationbenin
#> 5 lanationbenin
#> 6 lanationbenin