Main functions of collecting news from Côte Ivoir

Function for collecting news (title + content) from the media in Côte Ivoir.

ci_news_contents(start_d, end_d, press = "all")

Arguments

start_d
yyyy-mm-dd start date of collection (oldest date, eg: yesterday)

end_d
yyyy-mm-dd collection end date (newest date, eg: today)

press
specify the name of a press.

Value

a table containing the news (title + content + date publisher + press) collected

Examples

library(afnews)
dt_news <- ci_news_contents(start_d="2020-10-20", end_d="2020-10-25",press='all')

#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> Done!
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> Done!
#> Done!

head(dt_news)

#>   date_article
#> 1   2020-10-25
#> 2   2020-10-25
#> 3   2020-10-25
#> 4   2020-10-24
#> 5   2020-10-24
#> 6   2020-10-24
#>                                                                                                                                                                                                                        url_article
#> 1                                                                                              https://www.linfodrome.com/afrique-monde/62293-cameroun-les-hommes-en-armes-attaquent-une-ecole-primaire-plusieurs-morts-et-blesses
#> 2 https://www.linfodrome.com/afrique-monde/62288-guinee-cellou-dalein-diallo-et-ses-allies-rejettent-categoriquement-ces-faux-resultats-et-demandent-a-la-population-de-se-mobiliser-pour-exiger-le-respect-de-la-verite-des-urnes
#> 3                                                                                                           https://www.linfodrome.com/afrique-monde/62283-mali-l-armee-affirme-avoir-repris-farabougou-assiege-par-des-jihadistes
#> 4                                                                                                                      https://www.linfodrome.com/afrique-monde/62273-nigeria-plusieurs-entrepots-du-gouvernement-pilles-ce-samedi
#> 5                                                                                                                         https://www.linfodrome.com/afrique-monde/62272-cameroun-plusieurs-eleves-tues-dans-l-attaque-d-une-ecole
#> 6                                                                                         https://www.linfodrome.com/afrique-monde/62263-le-president-francais-emmanuel-macron-ordonne-la-condamnation-des-cyber-activistes-vereux
#>                                                                                   title_article
#> 1        Cameroun : Les hommes en armes attaquent une école primaire plusieurs morts et blessés
#> 2 Guinée : Cellou Dalein Diallo et ses alliés « rejettent catégoriquement ces faux résultats...
#> 3                    Mali : l’armée affirme avoir repris Farabougou, assiégé par des jihadistes
#> 4                                      Nigéria : Des entrepôts du gouvernement pillés ce samedi
#> 5                                   Cameroun : Plusieurs élèves tués dans l’attaque d’une école
#> 6     Le Président Français Emmanuel Macron ordonne la condamnation des cyber-activistes véreux
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        Content_article
#> 1 Le premier bilan officiel de l’attaque d’une école primaire à Kumba (Sud-ouest) samedi fait état de 6 morts et 13 blessés a déclaré le gouvernement camerounais.if (!window.AdButler){(function(){var s =document.createElement("script"); s.async = true; s.type ="text/javascript";s.src = 'https://servedbyadbutler.com/app.js';var n =document.getElementsByTagName("script")[0]; n.parentNode.insertBefore(s,n);}());}var AdButler = AdButler || {}; AdButler.ads = AdButler.ads || [];var abkw = window.abkw || '';var plc440701 = window.plc440701 || 0;document.write('<'+'div id="placement_440701_'+plc440701+'"></'+'div>');AdButler.ads.push({handler: function(opt){ AdButler.register(178368, 440701,[300,250], 'placement_440701_'+opt.place, opt); }, opt: { place:plc440701++, keywords: abkw, domain: 'servedbyadbutler.com',click:'CLICK_MACRO_PLACEHOLDER' }});Selon le gouvernement qui a dressé tard dans la nuit de samedi à dimanche, le premier bilan de l’attaque dans un établissement scolaire situé au quartier Fiango, arrondissement de Kumba 2 (Sud-Ouest), 6 élèves bien identifiés dont 1 garçon et 5 filles ont été tués, tous âgés de neuf à douze ans.René Emmanuel Sadi le ministre de la communication et porte-parole du gouvernement révèle également que 13 élèves ont été blessés soit 10 filles et 03 garçons, dont sept cas avérés préoccupants.Selon le gouvernement, l’attaque est l’œuvre d’un groupe de près d’une dizaine de terroristes, munis d’armes de guerre et constitués en véritable commando. Ledit groupe a fait irruption aux environs de 11 locales (10hTU), à bord de trois motocyclettes, dans l’enceinte du complexe scolaire privé dénommé « Mother Francisa International Bilingual Academy ».Par ailleurs, ledit établissement scolaire n’a lancé ses activités qu’en ce début d’année scolaire 2020/2021, « à l’insu des autorités administratives compétentes, et n’a pu bénéficier des mêmes mesures de protection que d’autres établissements scolaires du département de la Mémé », relève Yaoundé. Le gouvernement camerounais estime qu’aucune « cause au monde ne peut justifier ou légitimer de telles atteintes systématiques à la vie d’êtres humains, encore moins, à celle d’enfants innocents, d’élèves, ou d’étudiants, qui doivent aller à l’école, au collège, au lycée, à l’université, pour se former et être en mesure de se prendre en charge, et d’assumer dans le futur la relève de leurs aînés.» Le gouvernement camerounais a également appelé à la coopération internationale et à l’aide des pays amis qui hébergent ceux qui soutiennent, alimentent, financent le terrorisme au Cameroun et le climat d’insécurité dans les deux régions du Nord-ouest et du Sud-Ouest. Selon les autorités camerounaises, il ne fait pas de doute que l’attentat terroriste commis dans la ville de Kumba, traduit le « désarroi et la déroute des bandes armées sécessionnistes face à la détermination de la communauté éducative, de braver leurs menaces et leurs mots d’ordre, en décidant la reprise des classes dans leurs localités respectives.» Yaoundé pense que l’objectif des sécessionnistes est de briser la dynamique observée dans la reprise des classes dans le cadre de la rentrée scolaire 2020/2021 dans les régions du Nord-Ouest et du Sud-Ouest. Condamnations L’ensemble de la classe politique camerounaise a condamné un acte « terroriste d’une cruauté et d’une barbarie insoutenables » contre des jeunes innocents.Parfait KOFFI (function(d, s, id) {\n\t\t\t\t  var js, fjs = d.getElementsByTagName(s)[0];\n\t\t\t\t  if (d.getElementById(id)) {return;}\n\t\t\t\t  js = d.createElement(s); js.id = id;\n\t\t\t\t  js.src = "//connect.facebook.net/fr_FR/all.js#xfbml=1&appId=607728835948702";\n\t\t\t\t  fjs.parentNode.insertBefore(js, fjs);\n\t\t\t\t}(document, 'script', 'facebook-jssdk'));
#> 2                                                                                                                                                                                                                                                                                                                                               Au lendemain de la proclamation des résultats de l’élection présidentielle en Guinée, déclarant Alpha Condé vainqueur du scrutin avec 59,49%, l’Union des forces démocratiques de Guinée (UFDG) de Cellou Dalein Diallo et ses alliés « rejettent catégoriquement ces faux résultats et demandent à la population de se mobiliser pour exiger le respect de la vérité des urnes qui donne leur candidat vainqueur avec 53,84% ».if (!window.AdButler){(function(){var s =document.createElement("script"); s.async = true; s.type ="text/javascript";s.src = 'https://servedbyadbutler.com/app.js';var n =document.getElementsByTagName("script")[0]; n.parentNode.insertBefore(s,n);}());}var AdButler = AdButler || {}; AdButler.ads = AdButler.ads || [];var abkw = window.abkw || '';var plc440701 = window.plc440701 || 0;document.write('<'+'div id="placement_440701_'+plc440701+'"></'+'div>');AdButler.ads.push({handler: function(opt){ AdButler.register(178368, 440701,[300,250], 'placement_440701_'+opt.place, opt); }, opt: { place:plc440701++, keywords: abkw, domain: 'servedbyadbutler.com',click:'CLICK_MACRO_PLACEHOLDER' }});Alors qu’il est séquestré depuis plusieurs jours chez lui, le candidat autoproclamé vainqueur de la présidentielle du 18 octobre 2020, Cellou Dalein Diallo, a appelé, ce dimanche 25 octobre 2020, « la population à se mobiliser pour exercer le droit constitutionnel à manifester dans les rues et sur les places publiques pour dénoncer ce hold-up électoral en cours et pour exiger le respect de la vérité des urnes ». « Mes chers compatriotes. Comme vous le savez, la CENI a rendu public, hier les résultats provisoires globaux du scrutin présidentiel du 18 octobre 2020. Sans surprise, elle a validé les fraudes massives organisées après le scrutin, notamment lors des centralisations et a proclamé M.Alpha Condé vainqueur avec 59, 43%.L’UFDG et ses alliés de l’alliance nationale pour l’alternance et la démocratie rejettent catégoriquement ces faux résultats et demandent à la population de se mobiliser pour exiger le respect de la vérité des urnes qui donne leur candidat vainqueur avec 53,84%. L’UFDG et ses alliés demandent à la population de se mobiliser pour exercer le droit constitutionnel à manifester dans les rues et sur les places publiques pour dénoncer ce hold-up électoral en cours et pour exiger le respect de la vérité des urnes. L’UFDG et ses alliés demandent aux populations guinéennes de faire preuve d’une vigilance extrême pour éviter de tomber dans le piège d’Alpha Condé qui souhaite à tout prix, ethniciser le conflit pour opposer notre communauté et tirer, comme à son habitude, des bénéfices politiques.Comme vous le savez, ce conflit est strictement politique. Il oppose, d’une part, M.Alpha Condé et son clan, et d’autre part, l’écrasante majorité des Guinéens qui ne veulent pas du 3ème mandat et qui veulent défendre la vérité des urnes.Mes chers compatriotes, il n’y a pas d’alternative à la poursuite de notre combat. Cette élection, nous l’avons gagnée, haut les mains. Nous devons tout faire pour protéger, défendre notre victoire ».  (function(d, s, id) {\n\t\t\t\t  var js, fjs = d.getElementsByTagName(s)[0];\n\t\t\t\t  if (d.getElementById(id)) {return;}\n\t\t\t\t  js = d.createElement(s); js.id = id;\n\t\t\t\t  js.src = "//connect.facebook.net/fr_FR/all.js#xfbml=1&appId=607728835948702";\n\t\t\t\t  fjs.parentNode.insertBefore(js, fjs);\n\t\t\t\t}(document, 'script', 'facebook-jssdk'));
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  Les forces armées maliennes ont annoncé avoir repris le contrôle de Farabougou, un village du centre du pays qui était assiégé depuis deux semaines par des présumés jihadistes. Assimi Goïta, le vice-président de la transition, dirigeait les opérations, rapporte Jeune Afrique dans une publication du vendredi 23 octobre 2020.if (!window.AdButler){(function(){var s =document.createElement("script"); s.async = true; s.type ="text/javascript";s.src = 'https://servedbyadbutler.com/app.js';var n =document.getElementsByTagName("script")[0]; n.parentNode.insertBefore(s,n);}());}var AdButler = AdButler || {}; AdButler.ads = AdButler.ads || [];var abkw = window.abkw || '';var plc440701 = window.plc440701 || 0;document.write('<'+'div id="placement_440701_'+plc440701+'"></'+'div>');AdButler.ads.push({handler: function(opt){ AdButler.register(178368, 440701,[300,250], 'placement_440701_'+opt.place, opt); }, opt: { place:plc440701++, keywords: abkw, domain: 'servedbyadbutler.com',click:'CLICK_MACRO_PLACEHOLDER' }});Depuis deux semaines, la situation des habitants de Farabougou, dans le centre du Mali, était au centre de toutes les attentions. Encerclé par un groupe d’hommes armés présenté comme des jihadistes, ce village situé à environ 400 kilomètres au nord de Bamako était coupé du monde depuis le mardi 6 octobre, lorsqu’un groupe armé a fait irruption lors de la foire hebdomadaire et enlevé une vingtaine de personnes. Les assaillants ont ensuite pris position aux abords du village, bloquant toutes les voies d’accès.Le vendredi suivant, un groupe de villageois qui tentaient d’entrer dans Farabougou est tombé dans une embuscade. Cinq personnes ont été tuées et une quinzaine d’autres blessées. « Le village est coupé du pays par les terroristes qui ont bloqué toutes les voies d’accès », témoignait le dimanche 11 octobre le chef de village, Boukary Coulibaly. (function(d, s, id) {\n\t\t\t\t  var js, fjs = d.getElementsByTagName(s)[0];\n\t\t\t\t  if (d.getElementById(id)) {return;}\n\t\t\t\t  js = d.createElement(s); js.id = id;\n\t\t\t\t  js.src = "//connect.facebook.net/fr_FR/all.js#xfbml=1&appId=607728835948702";\n\t\t\t\t  fjs.parentNode.insertBefore(js, fjs);\n\t\t\t\t}(document, 'script', 'facebook-jssdk'));
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           Les pillages de masse ciblant des entrepôts des gouvernements ainsi que des entrepôts de nourriture se multiplient au Nigeria. Plusieurs milliers de personnes ont envahi et pillé, ce samedi, un entrepôt du gouvernement rempli de nourriture, à Jos, dans le centre du Nigeria, se plaignant "d'avoir faim".if (!window.AdButler){(function(){var s =document.createElement("script"); s.async = true; s.type ="text/javascript";s.src = 'https://servedbyadbutler.com/app.js';var n =document.getElementsByTagName("script")[0]; n.parentNode.insertBefore(s,n);}());}var AdButler = AdButler || {}; AdButler.ads = AdButler.ads || [];var abkw = window.abkw || '';var plc440701 = window.plc440701 || 0;document.write('<'+'div id="placement_440701_'+plc440701+'"></'+'div>');AdButler.ads.push({handler: function(opt){ AdButler.register(178368, 440701,[300,250], 'placement_440701_'+opt.place, opt); }, opt: { place:plc440701++, keywords: abkw, domain: 'servedbyadbutler.com',click:'CLICK_MACRO_PLACEHOLDER' }});Comme à Lagos et Ede (sud ouest) ces derniers jours, des milliers de personnes ont pillé des dizaines de milliers de sacs de nourriture qui auraient du leur être distribués pendant la période de confinement et de couvre-feu imposés pour contrôler l'épidémie de Covid-19.« Nous avions faim, et ils cachaient la nourriture ! », s’est indignée, une habitante de Jos. « Quel genre de gouvernement avons-nous ? Nous sommes faibles, beaucoup meurent de faim. » « Tous les prix ont augmenté. Comment faisons-nous pour survivre ? », a déploré un autre homme.Ces pillages surviennent après deux semaines de violences et de manifestations incontrôlables dans le pays. A Lagos, mégalopole économique de 20 millions d’habitants, la situation était plutôt calme ce samedi, et le couvre-feu strict imposé depuis mardi soir a été allégé.Le président Muhammadu Buhari, ancien général putschiste dans les années 1980, puis élu démocratiquement en 2015 et en 2019, a regretté les « nombreuses vies perdues », mais n’a donné aucun bilan officiel sur le nombre de victimes jusqu’à présent.Yves Leroy (function(d, s, id) {\n\t\t\t\t  var js, fjs = d.getElementsByTagName(s)[0];\n\t\t\t\t  if (d.getElementById(id)) {return;}\n\t\t\t\t  js = d.createElement(s); js.id = id;\n\t\t\t\t  js.src = "//connect.facebook.net/fr_FR/all.js#xfbml=1&appId=607728835948702";\n\t\t\t\t  fjs.parentNode.insertBefore(js, fjs);\n\t\t\t\t}(document, 'script', 'facebook-jssdk'));
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               Des hommes armés ont tué, ce samedi, plusieurs élèves lors de l’attaque d’une école à Fiango dans la commune de Kumba, selon l’ONU.if (!window.AdButler){(function(){var s =document.createElement("script"); s.async = true; s.type ="text/javascript";s.src = 'https://servedbyadbutler.com/app.js';var n =document.getElementsByTagName("script")[0]; n.parentNode.insertBefore(s,n);}());}var AdButler = AdButler || {}; AdButler.ads = AdButler.ads || [];var abkw = window.abkw || '';var plc440701 = window.plc440701 || 0;document.write('<'+'div id="placement_440701_'+plc440701+'"></'+'div>');AdButler.ads.push({handler: function(opt){ AdButler.register(178368, 440701,[300,250], 'placement_440701_'+opt.place, opt); }, opt: { place:plc440701++, keywords: abkw, domain: 'servedbyadbutler.com',click:'CLICK_MACRO_PLACEHOLDER' }});Au moins huit enfants ont été tués et douze ont été blessés i dans l’attaque de l’école de Kumba dans la région anglophone du sud-ouest du Cameroun.« Au moins huit enfants ont été tués par des coups de feu et des attaques à la machette. Douze autres ont été blessés et emmenés dans des hôpitaux locaux », a précisé dans un communiqué le bureau de la coordination des affaires humanitaires des Nations unies (OCHA) au Cameroun, sans préciser l’âge des victimes.Informés, les forces de sécurité camerounaises se sont rendues sur place après le carnage. Selon une source sécuritaire, il s’agirait d’une attaque des séparatistes anglophones. Ils s’opposent au retour de l’école dans les deux régions du Nord-ouest et du Sud-ouest.« L’attaque contre des écoliers est la pire atrocité depuis la reprise de l’année scolaire le 5 octobre 2020 dans laquelle plus d’étudiants se sont inscrits dans les régions du Nord-Ouest et du Sud-ouest que ces dernières années », a ajouté l’OCHA.Depuis près de trois ans, des groupes séparatistes et l’armée s’affrontent dans ces deux régions camerounaises où vit l’essentiel de la minorité anglophone, dont une partie s’estime marginalisée par la majorité francophone du pays.L’attaque n’a pas encore été revendiquée et les assaillants n’ont pas encore été identifiés par les autorités. Une réunion d’urgence était organisée samedi par le Premier ministre camerounais.Yves Leroy (function(d, s, id) {\n\t\t\t\t  var js, fjs = d.getElementsByTagName(s)[0];\n\t\t\t\t  if (d.getElementById(id)) {return;}\n\t\t\t\t  js = d.createElement(s); js.id = id;\n\t\t\t\t  js.src = "//connect.facebook.net/fr_FR/all.js#xfbml=1&appId=607728835948702";\n\t\t\t\t  fjs.parentNode.insertBefore(js, fjs);\n\t\t\t\t}(document, 'script', 'facebook-jssdk'));
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 Les réseaux sociaux au lieu de servir au raffermissement des liens, deviennent de plus en plus une plateforme d’expression des intentions obscures. Face à cette situation qui envenime la société, le président français veut sévir, en prenant des mesures contre tous ceux qui l’utilisent à des fins malsaines.if (!window.AdButler){(function(){var s =document.createElement("script"); s.async = true; s.type ="text/javascript";s.src = 'https://servedbyadbutler.com/app.js';var n =document.getElementsByTagName("script")[0]; n.parentNode.insertBefore(s,n);}());}var AdButler = AdButler || {}; AdButler.ads = AdButler.ads || [];var abkw = window.abkw || '';var plc440701 = window.plc440701 || 0;document.write('<'+'div id="placement_440701_'+plc440701+'"></'+'div>');AdButler.ads.push({handler: function(opt){ AdButler.register(178368, 440701,[300,250], 'placement_440701_'+opt.place, opt); }, opt: { place:plc440701++, keywords: abkw, domain: 'servedbyadbutler.com',click:'CLICK_MACRO_PLACEHOLDER' }});Le Président Emmanuel Macron, se dit préoccupé par les dérives sur les réseaux sociaux : « Il ne leur faut que quelques clics pour appeler à la haine, faire l'apologie du terrorisme, inciter à la violence. Il ne vous faut que quelques clics pour agir, signaler, faire condamner. Alors ne laissons pas faire ! ». Tel est le message qu’il a posté sur sa page Facebook.Son message, vu la menace du terrorisme dans le monde, il ne s’adresse pas seulement qu’aux français, mais aux habitants des autres pays. Quand on sait l’influence qu’a la France dans certains pays africains, son intervention y sera imminente.Carlos DOGBA (function(d, s, id) {\n\t\t\t\t  var js, fjs = d.getElementsByTagName(s)[0];\n\t\t\t\t  if (d.getElementById(id)) {return;}\n\t\t\t\t  js = d.createElement(s); js.id = id;\n\t\t\t\t  js.src = "//connect.facebook.net/fr_FR/all.js#xfbml=1&appId=607728835948702";\n\t\t\t\t  fjs.parentNode.insertBefore(js, fjs);\n\t\t\t\t}(document, 'script', 'facebook-jssdk'));
#>   press_article
#> 1    linfodrome
#> 2    linfodrome
#> 3    linfodrome
#> 4    linfodrome
#> 5    linfodrome
#> 6    linfodrome