Main functions of collecting news from Mali

Function for collecting news (title + content) from the media in Mali.

ml_news_contents(start_d, end_d, press = "all")

Arguments

start_d
yyyy-mm-dd start date of collection (oldest date, eg: yesterday)

end_d
yyyy-mm-dd collection end date (newest date, eg: today)

press
specify the name of a press.

Value

a table containing the news (title + content + date publisher + press) collected

Examples

library(afnews)
dt_news <- ml_news_contents(start_d="2020-10-20", end_d="2020-10-28", press='all')

#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> Done!
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> Done!

head(dt_news)

#>   date_article
#> 1   2020-10-28
#> 2   2020-10-28
#> 3   2020-10-28
#> 4   2020-10-26
#> 5   2020-10-26
#> 6   2020-10-26
#>                                                                                                                                                               url_article
#> 1                                                                            https://www.lesechos.ml/news/chronique-vive-lafrique-africaine-hommage-a-anna-diallo-kouyate
#> 2                                                                https://www.lesechos.ml/news/le-mali-veut-dialoguer-avec-ses-djihadistes-de-quoi-la-france-a-t-elle-peur
#> 3                                                                                                   https://www.lesechos.ml/news/le-m5-rfp-peut-il-se-remettre-sur-orbite
#> 4 https://www.lesechos.ml/news/mouvement-du-5-juin-rassemblement-des-forces-patriotiques-m5-rfp-declaration-liminaire-de-lassemblee-generale-du-m5-rfp-du-24-octobre-2020
#> 5                                                                 https://www.lesechos.ml/news/lassemblee-generale-le-m5-rfp-exige-a-limam-dicko-de-clarifier-sa-position
#> 6                                                                                                             https://www.lesechos.ml/news/editorial-ajuster-les-pendules
#>                                                                                                                                     title_article
#> 1                                                                           CHRONIQUE Vive l’Afrique  Africaine - Hommage à Anna Diallo Kouyaté !
#> 2                                                                   LE MALI VEUT DIALOGUER AVEC SES DJIHADISTES De quoi la France a-t-elle peur ?
#> 3                                                                                                      Le M5-RFP peut-il se remettre sur orbite ?
#> 4 Mouvement du 5 Juin – Rassemblement des Forces Patriotiques (M5-RFP) DECLARATION LIMINAIRE DE L’ASSEMBLEE GENERALE DU M5-RFP DU 24 OCTOBRE 2020
#> 5                                                                L’ASSEMBLEE GENERALE  Le M5-RFP exige  à l’imam Dicko de clarifier  sa position!
#> 6                                                                                                                EDITORIAL Ajuster les pendules !
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         Content_article
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  Tout y est imparfait. Me^me le plus parfait des imperfections. Toutes africainement africaines. Il n'y a de parfait que Celui qui sait naturellement tout parfaire. Notamment ce que l'Afrique tente difficilement d'ordonner. Par de´sordres incontournables. Cultive´s par re´flexe animal. La nature a horreur du vide. Et l'africain encore plus. En permanente transition. Nomadisme se´dentarite´. Rural urbanise´. Que^te de tombe e´clair. Toute gouvernance par terre. E´toiles filantes appre´cie´es. Repe´rages ce´lestes du chamelier pre´sidentiable. A la source du savoir e´ligible, Allah guide vers Sa crainte. Lui qui sait le cache´ et l'apparent. Afrique plus raisonnable que Descartes. Autre ente^- te´ caricaturiste islamophobe. Entre empereurs e´teints et allume´s qui tient qui ? Par la me`che de l'explosion thermonucle´aire plane´taire ! Allez savoir de´busquer le pirate au fond des oce´ans. Et le parcours d'une goutte de pluie dans le de´sert. Aussi bien ae´rienne que souterraine. Il en est de me^me de la vie de chaque mortel. A laquelle chaque a^me gou^tera. Et qu'est ce que l'a^me. Aimante ou amante. SOS dictionnaire de la folie. Sans ponctuation phrase´ologique. Ni appointement doctoral. Corps a` corps africain muscle´ squelette. De´veloppement par enveloppement des os fossiles. Capital arche´ologique des momies. Attention vivants de la mort. Touchez pas au festin sacrificiel des de´funts. Fils et filles de la jungle africaine urbanise´e. Fai^tes gaffe à la part destinée outre tombe Tous les lions sculpte´s ne sont pas e´dente´s. Pharaon a perdu ses griffes terroristes dans le Sinai¨ Jihadise´. Et Moise royalement installe Philipe Peugeot sur la peau du Lyons Club. Le di^ner est de´sormais servi aux domestiques a` table. Pliez les nattes pour la tapisserie de la mosque´e. A hauteur de la prosternation du lecteur de Coran assidu. A chacun sa bibliothe`que universelle. Ne touchez pas souille´s a` un seul verset du Dernier Porteur de la Parole d’ Allah (SWT). Vive l'Afrique Africaine. En hommage a` Anna Diallo Kouyate´ qui vient de retourner au Seigneur a` qui elle a bien appartenue. A Se´gou ce lundi de foire bien endeuille´e. Grande battante internationale USAID/Mali. Washington ne dira pas le contraire a` la Tribune New Yorkaise. Ici bas au dela` des murs de l'intole´rance police´s. Tre`s pieuse musulmane. Passerelle de joie pacifique entre deux mondes en feu. Aimable pre´sence d'esprit a` l'accueil. Forte capacite´ de concilier les inconciliables. Une sœur et e´pouse maternelle lumineuse nous a quitte´. Que faire de no^tre e´paisse obscurite´ malienne ? Prions sans rela^che pour tenir l’intenable en nous me^mes. Et surtout retenir une telle humanite´ hautement sensible. Paix a` son a^me ge´ne´reuse. Amin ! |IMAM PR HAMIDOU MAGASSA You can share this post!
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  Sur la gestion de la lutte contre le terrorisme, le Mali et la France ne regarde plus dans la direction. Si le gouvernement de transition dirigé par Moctar Ouane affiche de plus en plus un visage favorable à une possible négociation avec les Djihadistes maliens au regard de la lenteur des opérations militaires pour pacifier le pays, la France d’Emmanuel Macron s’oppose à cette méthode. Et elle ne le cache. Militairement unis au front dans la lutte contre le terrorisme, le Mali et la France sont désormais divisés politiquement depuis la volonté affichée par les autorités maliennes de transition de négocier les Djihadistes maliens conformément aux recommandations du Dialogue national inclusif et tel que demander par la CEDEAO qui incite les nouvelles autorités maliennes à appliquer le DNI. Mais, en visite de deux jours à Bamako, le ministre français des affaires étrangères, Jean Yves Le Drian, lors de la conférence de presse à la Primature et après l a signature de cinq conventions pour un montant estimé à plus de 93 milliards de F CFA, a fait connaitre publiquement son opposition à cette option du Peuple malien et que le gouvernement de transition n’est autre qu’un simple exécutant. Pour le ministre français des affaires étrangères, Jean Yves Le Drian, au Mali, la négociation n’est possible qu’avec les groupes armés signataires de l’Accord pour la paix et la réconciliation issu des pourparlers d’Alger. « Il y a des accords de paix. Il faut mettre en œuvre les accords de paix… Ils ont été validés par un certain nombre de signataires, dont des groupes armés signataires. Et puis, il y a les groupes terroristes qui n’ont pas signé les accords de paix. Les choses sont simples », a indiqué le chef de la diplomatie française qui veut visiblement dicter sa volonté au Peuple souverain du Mali. Face à cette dictature politique, il a droit à un petit rappel de la part du Premier ministre malien, Moctar Ouane, lui aussi pas prêt à jeter l’espoir des Maliens à l’eau. En effet, Moctar Ouane n’exclut pas l’éventualité d’une négociation avec les terroristes : « Je crois également utile sur cette question de me faire l’écho des conclusions du Dialogue national inclusif qui a eu lieu chez nous et qui a très clairement indiqué la nécessité d’une offre de dialogue avec ces groupes armés. Je pense qu’il faut voir en cela une opportunité d’engager de vastes discussions», a rappelé le Chef du gouvernement malien. Pour rappel, l’idée de dialoguer avec les Djihadistes n’est pas une création de Moctar Ouane. Ce projet est plus vieux que sa nomination à la Primature. C’est un projet du Peuple souverain du Mali dont la modification dépasse le pouvoir de Moctar Ouane et son chef Bah N’Daw  réunis. La négociation avec les Djihadistes maliens, une des recommandations du Dialogue national inclusif (DNI) tenu en décembre dernier à Bamako, était aussi une des conclusions de la conférence d’entente nationale tenue du 27 mars au 2 avril 2017. C’est aussi un souhait exprimé à maintes reprises par Tiébilé Dramé, ancien ministre des Affaires étrangères du Mali, et l’imam Mahmoud Dicko, ancien président du Haut Conseil Islamique du Mali(HCI). Mieux, depuis la tenue du DNI, la CEDEAO ne cesse d’inciter les autorités maliennes à son application stricte et rapide. Même le Secrétaire général de l’ONU, Antonio Guterres, semble adhéré à cette idée du Peuple malien de dialoguer avec « certains groupes extrémistes ». « Il y aura des groupes avec lesquels on pourra parler, et qui auront intérêt à s’engager dans ce dialogue pour devenir des acteurs politiques dans le futur », déclarait le Secrétaire général de l’ONU lors d’un entretien donné, le 19 octobre dernier, au quotidien Le Monde. Alors question : de quoi la France a-t-elle peur dans son opposition en solo à toute idée de dialoguer avec les Djihadistes maliens ? Les jours à venir seront plus édifiants ! Youssouf Z KEITA  You can share this post!
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   Leur mouvement ressemble de plus en plus à un vaisseau qui se vide de son équipage. Pourtant, les M5-RFPéistes, qui ont du mal à renouveler leurs objectifs, affirment être toujours d’attaque. Laissons à l’avenir proche le soin de lever le voile sur la réalité. La politique est décidément le royaume où le grand paradoxe est la règle, et où la fraternité sans lendemains est le principe. Vous vous mobilisez contre un ennemi commun, vous vous jurez serment d’amour et de fidélité, puis vous formez une alliance sacrée pour le terrasser et le déloger de la citadelle de Koulouba. Mais, une fois cet ennemi vaincu, que vous arrive-t-il donc ?!!!Oh ! Voilà que vous êtes vous-mêmes surpris de ressentir au plus profond de vous comme un sentiment de vide.  Eh, oui ! Aussi inexplicable que cela paraisse, vous commencez à vous ennuyer car votre ennemi-là vous manque. N’est-ce pas face à lui que vous étiez parvenus à incarner ce qu’il y a de plus vertueux dans l’exercice de la politique ? N’est-ce pas contre lui que vous étiez enfin parvenus à redorer votre réputation auprès de la population ? N’est-ce pas en s’attaquant à lui que vous étiez devenus les chouchous du peuple en émoi ? Du coup, sans lui, vous êtes en manque de punch, vous êtes en manque de qui sur qui cogner. Alors, votre enthousiasme collectif s’évapore lentement mais sûrement, et vous avez du mal à prendre goût aux choses, sans le bon vieil ennemi à combattre. Rendus alors nostalgiques du féroce combat qui vous plaçait en face de votre ennemi intime, ne pouvant rembobiner les épisodes où sans cesse vous lui donniez des coups et en preniez de lui ; vous éprouvez le plus grand mal à renouveler vos ambitions et à trouver votre place dans l’arène. Malgré vous, vous êtes déchirés : les uns choisissent de se mettre en sommeil ; les autres nourrissent comme un besoin de rompre l’alliance et de suivre chacun sa propre étoile. Sans IBK à combattre, le M5-RFP (ou plutôt ce qu’il en reste) semble se chercher une seconde vie. Les ténors du mouvement (ou plutôt ceux qui s’accrochent aux vestiges) promettent de rallumer la flamme et de réveiller le Kraken. Mais, réactiver le feu pour quoi faire ? Tirer le Kraken de son repos pour quel nouveau noble but ? Tel est le délicat questionnement auquel Choguel Kokala Maïga et les autres grosses planètes de la galaxie M5 vont devoir répondre. Leur tâche s’annonce pourtant bien difficile car deux vents contraires menacent de contrarier davantage la trajectoire déjà comprise du vaisseau M5. Le premier vent est celui de « la dure réalité du quotidien » dont le souffle torride enveloppe les Maliens au point de les rendre peu réceptifs à un nouveau discours qui prône le renverse-tout. La gestion calamiteuse du trio IBK-RPM-ADEMA a été si pernicieuse que les citoyens en ressentent maintenant plus que jamais les effets de la manière la plus insupportable. Les Maliens suffoquent. Ils sont asphyxiés par la précarité. Les Maliens halètent. Ils sont étranglés par l’amoindrissement de leurs revenus. Les Maliens doutent de l’avenir. Ils en sont réduits à racler les fonds de tiroirs pour affronter un aujourd’hui assombri par des années de mal gouvernance. Dans ce contexte, le peuple anxieux et déprimé ne va-t-il pas privilégier "Le Laboureur et ses enfants" de Jean de La Fontaine au lieu de « la vision rouge de la révolution » décrite par Emile Zola dans "GERMINAL" ? Le second vent, mortifère celui-là, qui risque de casser définitivement le gouvernail du M5, est celui de la division et des défections qui se multiplient dans le cockpit. Des figures de proue du comité stratégique ont beau chercher à nier ou à minimiser l’évidence, les faits sont là. Le Mouvement du 5 juin subit bel et bien le départ de certains de ses cadres les plus influents. Qui peut oser confirmer que la CMAS et l’imprévisible Issa Kaou Djim comptent toujours parmi les membres de l’équipage ? Qui peut nier qu’Oumar Mariko et ses lieutenants du MDP ont rangé leurs cartes du M5 au placard ? On le voit. C’est incontestable. Le M5, regroupement de circonstance, a commencé de s’étioler aussitôt que son meilleur ennemi, IBK, a été évincé de l’arène. A présent, si les M5-RFPéistes veulent pouvoir se remettre sur orbite et incarner l’espoir d’ « un Mali nouveau », ils doivent d’abord se sauver d’eux-mêmes. En clair, un : arrêter de paraître seulement comme un ouragan de la contestation à l’affût et prêt à bondir au moindre faux pas des autorités en place ; deux : se projeter déjà dans l’après-Transition en concoctant patiemment un projet commun de société capable d’aimanter le vote de la majorité des Maliens. Ils pourront alors clamer aux citoyens : « Le changement que nous vous avions promis, c’est maintenant ! » Mohamed Meba TEMBELY You can share this post!
#> 4 Militantes et militants du M5-RFP ou plus exactement militantes et militants d’un Mali debout, d’un Mali digne pour l’édification duquel notre mouvement a tout donné, jusqu’au sang de patriotes lâchement assassinés à Sikasso, Kayes et ici à Bamako. Nous avons fait le serment que justice leur sera rendue. Nous honorerons notre serment envers eux. Et envers toutes les victimes militaires et civiles des crises que nous vivons. Pour notre honneur. Pour l’honneur du Mali. Militantes et militants du M5-RFP, Nous voici donc réunis, depuis la démission de l’ancien président de la République, pour notre 1ère Assemblée générale qui regroupe les responsables de nos principales entités faîtières. Le Comité stratégique, comme à chaque moment important de la vie de notre mouvement, souhaite vous rendre compte et surtout vous écouter. Militantes et militants du M5-RFP, Nous devons tout d’abord vous dire toute notre fierté et vous transmettre les milliers de message de remerciement et de félicitations, pour ce que vous avez entrepris et réussi sous le regard admiratif de l’Afrique et au-delà du monde entier. Peu de gens croyaient que les mains nues, dans le respect de la loi que vous auriez obtenu le départ de l’ancien président. Voilà que IBK est parti : • Merci au peuple malien qui nous a fait confiance en participant massivement au soulèvement populaire contre l’ancien régime corrompu, oligarchique et ploutocratique ; • Gloire aux militants et sympathisants du M5-RFP avec une mention spéciale les jeunes et aux femmes pour leur formidable engagement et les sacrifices consentis ; • Hommages aux martyrs et à toutes les victimes de cette lutte : leur sacrifice ne sera pas vain. Militantes et militants du M5-RFP, Après la démission de l’ancien président, nous avons pensé et d’ailleurs nous continuons à penser que la condition sine qua non de la réussite de la Transition était une saine et loyale collaboration entre les deux acteurs du changement, entre les deux légitimités incarnées par le M5-RFP et le CNSP. Et croyez-nous chers militantes et militantes, nous n’avons été dupes de rien et rien ne nous a surpris. En effet : – en quoi peuvent surprendre des gens qui ne prononcent le mot M5-RFP que lorsqu’ils sont contraints et forcés ? – en quoi peuvent surprendre des gens qui, dès le lendemain du 18 août ont renoué avec les pratiques que nous combattions la veille seulement ? – en quoi peuvent surprendre des gens qui ont très vite fait la preuve que le Mali n’était pas leur souci et qu’ils sont plutôt en quête de privilèges. Mais il nous fallait laisser les maliens, voir, entendre, constater et comprendre par eux-mêmes ce qui se tramait contre eux. Ainsi, dans le seul but de marginaliser le M5-RFP et d’accaparer tout le pouvoir pour ne pas avoir en face une véritable légitimité qui ne serait ni otage ni impuissant : – tous les engagements pris, même sous serment d’officier, ont été violés ; – les conclusions des concertations sur la Transition ont été falsifiées ; – le Président et le Premier de Transition ont été choisis par malice et fraude ; – les textes régissant la Transition n’ont respecté aucune règle. Ce qui ne pouvait manquer d’arriver arriva et se constatera tous les jours un peu plus : en quelques semaines l’espoir s’est transformé en défiance, la confiance en déception et la Transition s’achemine vers une inéluctable impasse. Mais chers militantes et militants, là est notre LIGNE ROUGE. Nous ne laisserons personne empêcher l’émergence d’un Mali nouveau. Nous n’accepterons pas la restauration en cours. Nous dénoncerons et combattrons les complots de plus en plus manifestes contre notre peuple. En un mot nous ne laisserons personne perpétuer le régime de IBK sans IBK. Militantes et militants du M5-RFP, Comme hier, malgré tout ce qui a été annoncé et souhaité sur le caractère hétéroclite de notre mouvement nous restons unis. Mais comme hier, nous ne disons pas que nous sommes monolithiques. C’est le lieu pour nous de prendre acte de l’annonce de l’Imam Mahmoud Dicko de se consacrer désormais à sa mission de guide religieux. Il a d’ailleurs décidé de s’exprimer lui-même sur ce sujet au moment opportun. Comme hier nous sommes déterminés à obtenir le changement pour la Refondation du Mali. Et comme hier nous y arriverons en imposant une Transition de rupture. Oui, dans les organes et institutions ou en dehors nous imposerons une Transition de rupture avec l’ordre ancien est ses pratiques assassines contre notre Peuple. Nous poserons les bases solides d’une refondation du Mali basée sur des objectifs clairs dont : 1). Ancrage durable du Mali dans la paix, la sécurité, l’unité nationale, le recouvrement de l’intégrité du territoire national et de la souveraineté du Mali ; 2). Gouvernance vertueuse : fondée sur une lutte implacable contre la corruption, le traitement diligent des rapports de vérification (ceux disponibles et ceux élaborés pendant la transition) ; l’audit des finances publiques (dont celui de la dette) ,des institutions de la république , des ministères et grandes structures publiques et de l’aliénation du patrimoine de l’Etat ; 3). Consolidation de notre système démocratique : réformes politiques et institutionnelles ; organisations d’élections crédibles ; 4). Consolidation de l’Etat de droit : procès crime de sang ; reforme et renforcement des capacités de la justice ; accès à et bonne distribution de la justice 5). Amélioration des conditions de vie des populations par, entre autres mesures, le retour de l’Etat sur l’ensemble du territoire national, la relance de l’économie et le soutien aux secteurs sinistrés par le Covid-19. Dans cette optique nous pèserons de toutes nos forces sur le processus de gouvernance par des actions ciblées : – Soutien aux mesures des autorités de transition favorables à la refondation du Mali ; – Contestation vigoureuse des mesures de maintien ou de renforcement des mauvaises pratiques de gouvernance non favorables à la refondation ; – Exercice de pressions pour la prise de mesures porteuses de refondation. L’idéal aurait été que les acteurs du changement, dans le cadre des organes de la Transition, mettent en œuvre le changement. Oui, nous avons voulu être les principaux responsables de la transition pour imprimer notre vision à la marche des affaires. Sans complexe et en toute légitimité. Dans cette optique et toujours en toute légitimité le M5-RFP réclame, comme un dû, la présidence et le quart des membres du Conseil national de Transition pour encore une fois se donner les moyens politiques et institutionnels du changement et de la refondation. En toute hypothèse, militantes et militants du M5-RFP le Comité stratégique, au nom de l’ensemble des composantes du Mouvement, invite les maliennes et maliens à rester mobilisés et déterminés pour une rupture véritable avec l’ancien système qui perdure et pour un changement réel par la Refondation du Mali. Nous vous remercions de votre attention.  Bamako, le 24 octobre 2020 Pour le Comité stratégique du M5-RFP Choguel K. Maïga  You can share this post!
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            Deux mois après la chute de l’ancien Président, le Mouvement du 5 juin (Rassemblement des Forces Patriotiques) a tenu, samedi 24 octobre, sa première Assemblée générale, dans l’enceinte de la CSTM. Elle a enregistré la présence de Dr Choguel  Maiga, Modibo Sidibé, Mme Sy Kadiatou Sow, Me Mountaga Tall… et  de nombreux  militants et sympathisants du Mouvement. Le Mouvement du 5 juin qui n’avait pas organisé une grande rencontre, depuis la démission de l’ancien chef de l’Etat, a tenu, le samedi dernier, sa toute première Assemblée générale. Selon le communiqué rendu public, le M5-RFP a pensé et continue à penser que la condition sine qua non de la réussite de la transition était une saine et loyale collaboration entre les deux acteurs du changement, entre les deux légitimités incarnées par le M5-RFP et le CNSP. « En quoi peuvent surprendre des gens qui, dès lendemain du 18 août, ont renoué avec les pratiques que nous combattions la veille ? », s’interrogent les dirigeants du Mouvement. Selon eux, il fallait laisser les maliens, voir, entendre, constater et comprendre par eux-mêmes ce qui  se tramait contre eux. Ainsi, le comité stratégique accuse le Cnsp d’avoir marginalisé le M5 et de s’emparer de tout le pouvoir pour ne pas, disent-ils, « avoir en face une véritable légitimité qui ne serait ni otage ni impuissante ». Pour le président du M5-RFP, le Dr Choguel Maiga, tous les engagements pris sous serment d’officier ont été violés.« Les concertations sur la transition ont été falsifiées, le président et le premier ministre de la transition on été choisis par malice et fraude, les textes régissant la transition n’ont respecté aucune règle », a-t-il dénoncé. Le M5-RFP promet de conjuguer ses efforts pour poursuivre son combat. « Comme hier, malgré tout ce qui a été annoncé et souhaité sur le caractère hétéroclite de notre mouvement, nous restons unis, mais, précise-t-il, comme hier, nous ne disons pas que nous sommes monolithiques ». Le mouvement  a pris acte de l’annonce de l’Imam Dicko qui dit vouloir se retirer du combat pour se consacrer désormais à sa mission de guide religieux. « L’imam Dicko a décidé de s’exprimer lui-même sur ce sujet au moment opportun », a fait savoir Imam Oumarou Diarra. Un intervenant prenant la parole à son tour s’oppose : « On doit savoir la position de l’Imam Dicko. Il refuse de parler alors que nos militants ont été tués devant sa maison et dans sa mosquée. Quand son domicile, explique-t-il, a été attaqué, moi-même, j’ai fait appel à mes éléments pour venir le sauver ». Dans sa déclaration, le M5-RFP réclame la présidence du Conseil National de la Transition (CNT). « Comme un dû, demandons la présidence et le quart des membres du Conseil National de la Transition pour encore une fois se donner les moyens politiques et institutionnels du changement et de la refondation » ont-ils sollicité. Seydou FANE You can share this post!
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 Ne dit -on pas au Sêno que c'est "une oreille bénie qui entend les conseils "? Que de conseils prodigués au Président IBK et à son régime? Les thuriféraires et leur mentor ont considéré ces conseils au mieux comme venant de citoyens " affamés " qui cherchent des faveurs et au pire comme venant de "petites gens" des " gens rien " des "hasidiw".La suite est connue. Les laudateurs se sont terrés ou ont retourné leurs vestes. Le peuple malien qui cherche à sortir de cet infernal chaos dans lequel il baigne depuis plusieurs années,fonde un immense espoir sur la Transition. Or les responsables de la Transition par leurs gestes quotidiens laissent nombre de maliens dubitatifs. Les nombreuses revendications sociales et syndicales qui sont d'une importance et d'une acuité extrême doivent être discutées, négociées et échelonnées au besoin La mise en place du CNT.,l'organe législatif de la Transition, doit être une occasion de réunir pendant une semaine au moins des représentants des différentes couches socioprofessionnelles pour choisir les membres du CNT mais surtout négocier des moratoires compte de la situation financière catastrophique de notre pays Les préfets et autres agents municipaux qui sont la chevillé ouvrière des différentes élections ont déjà battu le pavé malgré l'existence d'une ordonnance du 30 Juillet 2018, pour exiger une amélioration de leurs conditions sécuritaires et de travail. Certains agents de la santé sont aussi déjà en grève alors les magistrats seraient en embuscade. Gouverner c'est prévoir mieux gouverner c'est anticiper. Il faut "attraper la sauterelle alors qu'elle a les ailes encore lourdes " dit-on .La Transition qui semble s'aliéner toutes forces politiques ou sociales résisterait difficilement aux différents soubresauts. Il est encore temps d'ajuster les pendules. Qu'Allah veille sur notre Mali. Hamidou Konaté  You can share this post!
#>   press_article
#> 1      lesechos
#> 2      lesechos
#> 3      lesechos
#> 4      lesechos
#> 5      lesechos
#> 6      lesechos