Main functions of collecting information on Senegal

Function for collecting news (title + content) from the media in Burkina Faso.

sn_news_contents(start_d, end_d, press = "all")

Arguments

start_d
yyyy-mm-dd start date of collection (oldest date, eg: yesterday)

end_d
yyyy-mm-dd collection end date (newest date, eg: today)

press
specify the name of a press.

Value

a table containing the news (title + content + date publisher + press) collected

Examples

library(afnews)
dt_news <- sn_news_contents(start_d="2020-10-28", end_d="2020-11-02", press='all')

#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> Done!

#> Warning: No article publish in this period !

#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> Done!

head(dt_news)

#>   date_article
#> 1   2020-11-02
#> 2   2020-10-29
#>                                                                          url_article
#> 1        http://lesoleil.sn/nouveau-gouvernement-plus-quun-jeu-de-chaises-musicales/
#> 2 http://lesoleil.sn/infographie-des-cinq-gouvernements-sous-macky-sall-depuis-2012/
#>                                                                title_article
#> 1                 Nouveau gouvernement: Plus qu’un jeu de chaises musicales…
#> 2 Infographie I Focus sur les cinq gouvernements sous Macky Sall depuis 2012
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        Content_article
#> 1 Stabilité et ouverture, jeunesse et relance de l’économie : le Gouvernement nommé hier est plus qu’un turn-over, c’est un vrai changement d’équipe. On y note d’importants départs tout comme il consacre la nouvelle alliance avec l’homme arrivé deuxième lors de la dernière présidentielle. Il va dérouler son action dans le cadre d’une majorité présidentielle élargie avec, comme catalyseur, la nomination de l’ancien Premier ministre Idrissa Seck à la présidence du Conseil économique, social et environnemental (Cese) alors que son parti, Rewmi, a obtenu deux sièges en Conseil des ministres. Les décisions présidentielles vont donc au-delà de l’exécutif, car touchant une institution. Ce remaniement ministériel n’est pas qu’un simple jeu de chaises musicales, c’est un nouvel orchestre qui monte sur scène, même si, comme à son habitude, le Président de la République n’a pas fait table rase de ses choix précédents dont celui de certains ministres présents au Gouvernement depuis son accession au pouvoir en 2012. Ce remaniement est l’expression de la volonté présidentielle de (re) mettre les défis de l’heure au cœur de l’activité gouvernementale et de constituer un collège composé de responsables exclusivement -mais alors exclusivement- consacrés à l’atteinte des objectifs qu’il a définis dans le programme proposé aux Sénégalais en 2019. Au fond, le cadre conceptuel, la vision à long terme et la « Politique », pour tout dire, n’ont pas varié. Nous sommes dans la phase II du Plan Sénégal émergent (Pse). Il y a eu certes des ajustements illustrés par le Programme de résilience économique et sociale pour faire face à la pandémie dont les effets dans certains secteurs ont été dévastateurs. Il y a aussi la création du Force Covid-19 doté de 1000 milliards de FCfa ; ce qui a différé le lancement et l’accélération de projets stratégiques. OBJECTIFS Mais, la vision globale demeure. Ce nouveau Gouvernement a comme feuille de route « Les 5 initiatives majeures » dans la séquence de la seconde phase du Plan Sénégal émergent (Pse). « Avec 44 % de nos compatriotes qui ont moins de 15 ans, l’éducation et la formation professionnelle se présentent comme axes de progrès pour promouvoir davantage l’employabilité des jeunes », avait souligné Macky Sall en présentant son programme à ses compatriotes en fin 2011. L’économie sociale et solidaire, pour compléter les services financiers déjà offerts par la Der (Délégation générale à l’entrepreneuriat rapide), par un nouveau package de services non-financiers incluant la formation et l’encadrement des bénéficiaires  « afin que les demandeurs d’emplois d’aujourd’hui soient les créateurs d’emplois de demain » ; l’économie numérique inclusive ; le Pse vert, enfin, la cinquième, une initiative d’industrialisation à grande échelle, « clé de la transformation structurelle, économique et sociale que nous portons à travers le Pse », complètent le programme du Gouvernement. Ces cinq initiatives majeures sont complétées sur la même séquence 2019-2024 par trois programmes sectoriels à l’échelle nationale. Il s’agit du programme zéro bidonville, du programme zéro déchet, du programme villes créatives « qui replace les industries culturelles au cœur de la Cité ». Voilà les objectifs fixés au nouveau Gouvernement. Élu et bien élu au premier tour, en 2019, avec plus de 58,27 % des suffrages (66,23 de taux de participation), Macky Sall a vu son élan lesté, depuis un an, par la crise du coronavirus, mais face à laquelle bien des leçons ont été tirées et même des satisfactions notées, à voir la qualité de la riposte. Sur le plan politique, il lui a surtout été empêché de mobiliser l’opinion autour de l’essentiel avec un débat pollué par la question des mandats et les ambitions présidentielles nées dans son propre camp. C’est un secret de polichinelle, les personnalités dont l’agenda paraissait peu compatible avec la participation, à des niveaux stratégiques, à un Gouvernement censé être uniquement voué à la réussite du bilan de Macky Sall, sont parties. Mais, c’est là un gène du chromosome politique sénégalais. Les acteurs politiques, les médias et la société civile, tous sont englués dans les conjectures autour du fait électoral. À travers des commissions, des concertations nationales, des assises, l’élite politique a privilégié la gestion des contradictions autour des élections au détriment de l’animation du clivage majorité/opposition sur la base de divergences sur des choix de société et des politiques publiques (système de santé, programmes sociaux, fiscalité, programmes scolaires, Code de la famille, décentralisation, service militaire obligatoire, peine de mort…) Les urnes « encombrent » la démocratie sénégalaise même si elles en sont le « Graal ». On ne peut pas sortir d’une présidentielle sans contestation pour tomber sur la polémique relative à la possibilité pour Macky Sall de se présenter en 2024. Une société politique s’anime d’abord et avant tout autour de thèmes qui touchent les préoccupations des citoyens. Naturellement, les positions tranchées autour des mandats du Président de la République, les ambitions des uns et des autres pour la magistrature suprême et la longue culture électorale sénégalaise font que les spéculations autour de 2024 soient attractives. Toutefois, c’est prématuré. Et le pays a d’autres priorités.
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       D’avril 2012 à octobre 2020, cinq gouvernements ont été formés sous la présidence de Macky Sall. Notre infographie sur ces différentes équipes. Le gouvernement Abdoul Mbaye est nommé le 4 avril 2012 à la suite de la victoire du candidat de la coalition Benno Bokk Yakkar, Macky Sall. Le gouvernement est resserré à 25 ministres. Un remaniement intervient le 29 octobre 2012. Il sanctionne le départ de sept ministres, dont des proches collaborateurs de Macky Sall, Alioune Badara Cissé (Affaires étrangères), Mbaye Ndiaye (Intérieur), Mor Ngom (Transports) et Aly Coto Ndiaye (Jeunesse), ainsi qu’Ibrahima Sall (Éducation), Mata Sy Diallo (Commerce), et Abou Lô (Communication). Douze nouveaux ministres sont nommés, portant à trente le nombre de membres du gouvernement. Deux des sept remerciés sont nommés auprès du président de la République : Mbaye Ndiaye, comme ministre d’État auprès du président de la République et Mor Ngom comme directeur de cabinet avec rang de ministre d’État.  Aminata Toure est la deuxième personnalité à occuper la fonction de Premier ministre sous la présidence de Macky Sall. Elle est à la tête du gouvernement de la République du Sénégal entre le 2 septembre 2013 et le 6 juillet 2014. Elle a dirigé un gouvernement composé de 31 ministres et un ministre délégué, alors que l’équipe précédente comptait 30 membres. La composition s’est faite majoritairement à partir des ministres sortants du gouvernement Mbaye, auxquels viennent s’ajouter huit nouveaux ministres. Mahammed Boun Abdallah Dionne a dirigé à partir du 6 juillet 2014, le troisième gouvernement sous la présidence de Macky Sall. Ce gouvernement est composé de 39 membres, dont 30 ministres, 3 ministres délégués et 6 secrétaires d’État, alors que le précédent gouvernement comptait 32 membres. Sept femmes font partie du gouvernement. Le 22 juin 2015, un remaniement intervient : Omar Youm quitte le gouvernement, pour devenir Directeur de cabinet du président de la République, et est remplacé au ministère de la Gouvernance locale, du Développement et de l’Aménagement du territoire par Abdoulaye Diouf Sarr, qui laisse son portefeuille de l’Energie à Thierno Alassane Sall, ancien ministre des Infrastructures, du Transport terrestre et du Désenclavement dans les gouvernements Mbaye et Touré. Le gouvernement démissionne le 5 septembre 2017, pour laisser place au Gouvernement Dionne II le 7 septembre.  Le gouvernement Dionne II est formé le 6 septembre 2017. Il s’agit du quatrième gouvernement sous la présidence de Macky Sall. Il est composé de 39 membres, dont 35 ministres, 4 ministres délégués. Après la démission du premier gouvernement Dionne, le 5 septembre 2017, le président Macky Sall reconduit Dionne comme Premier ministre, par décret n° 2017-1531 du 6 septembre 2017, et nomme le lendemain un nouveau gouvernement, où se retrouve l’essentiel des personnalités du cabinet précédent, à l’exception de Mankeur Ndiaye et d’Awa Marie Coll Seck, nommée toutefois quelques jours plus tard ministre d’État auprès du président de la République. Le Président de la République, Macky Sall a nommé, le 07 avril  2019, un gouvernement composé de 32 ministres, 3 secrétaires d’État, contre 39 ministres dans le précédent gouvernement. Le nouvel attelage gouvernemental a été renouvelé à 50% avec 25% de femmes. Ce Gouvernement était dirigé par le Premier ministre, ministre d’Etat et Secrétaire général de la Présidence, Mahammad Boun Abdalah Dionne.
#>   press_article
#> 1      lesoleil
#> 2      lesoleil