Main functions of collecting news from Burkina Faso

Function for collecting news (title + content) from the media in Burkina Faso.

bf_news_contents(start_d, end_d, press = "all")

Arguments

start_d
yyyy-mm-dd start date of collection (oldest date, eg: yesterday)

end_d
yyyy-mm-dd collection end date (newest date, eg: today)

press
specify the name of a press.

Value

a table containing the news (title + content + date publisher + press) collected

Examples

library(afnews)
dt_news <- bf_news_contents(start_d="2020-10-20", end_d="2020-10-28", press='all')

#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> Done!
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> Done!

head(dt_news)

#>   date_article
#> 1   2020-10-28
#> 2   2020-10-27
#> 3   2020-10-27
#> 4   2020-10-25
#> 5   2020-10-22
#> 6   2020-10-28
#>                                                                                                                                                url_article
#> 1                                                                                                                 http://lefaso.net/spip.php?article100293
#> 2                                                                                                                 http://lefaso.net/spip.php?article100259
#> 3                                                                                                                 http://lefaso.net/spip.php?article100244
#> 4                                                                                                                 http://lefaso.net/spip.php?article100199
#> 5                                                                                                                 http://lefaso.net/spip.php?article100142
#> 6 https://www.burkina24.com/2020/10/28/lancement-de-lincubateur-sira-labs-a-bobo-dioulasso-des-opportunites-de-creation-dentreprises-numeriques-pour-tous/
#>                                                                                                                                        title_article
#> 1                                                            Coopération : Le Japon appuie la réalisation de quatre forages dans la commune de Toécé
#> 2                                                                                     Coopération : Suppression de visa entre le Burkina et le Maroc
#> 3 Sécurité : L’Union européenne et le Burkina Faso passent en revue les actions de coopération menées dans le cadre de la lutte contre le terrorisme
#> 4                                          Bobo-Dioulasso : Le ministre de l’Intégration africaine à l’écoute des communautés vivant au Burkina Faso
#> 5                             Intégration africaine : Des communautés vivant au Burkina Faso outillées sur leurs droits et devoirs dans les Cascades
#> 6                             Lancement de l’incubateur Sira Labs à Bobo-Dioulasso : Des opportunités de création d’entreprises numériques pour tous
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  Content_article
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            Après le soutien du Japon à la commune de Tanghin-Dassouri pour la construction de lampadaires au profit des élèves, c’est autour de la commune de Toécé de bénéficier de l’appui du peuple nippon.  Le Japon a en effet mis à la disposition de l’Association conscience et action pour le développement (ASCADE), la somme de 29 800 000 F CFA. Cette somme va servir à la réalisation de quatre forages positifs dans à Toécé.\nCes forages viendront accroître l’accès à l’eau potable au profit des populations, en particulier des élèves. L’accord de financement a été signé par le responsable de l’ASCADE, Alain Nikiéma, et l’ambassadeur du Japon au Burkina Faso, Masaaki Kato.\nRépondant aux mots de remerciement d’Alain Nikiéma, Masaaki Kato a déclaré : « Le Japon est un ami du Burkina. Ces forages vont permettre de soulager 1 200 élèves qui ne seront plus obligés de parcourir environ 2 km pour avoir accès à l’eau potable. On pourra lutter efficacement contre les maladies d’origine hydrique ».\nE.K.S.\nLefaso.net
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         Le 23 octobre 2020, Alpha Barry, ministre des Affaires étrangères et de la coopération, était au Maroc. Ce jour, il a inauguré, en compagnie de son homologue marocain, Nasser Bourita, le Consulat général du Burkina à Dakhla. Pour le ministre Bourita, « l’inauguration de cette représentation consulaire contribuera au renforcement de la coopération économique entre les deux pays, en cohérence avec la place de choix que le Souverain accorde à la ville de Dakhla, en tant que locomotive économique, étant donné qu’elle constitue un véritable trait d’union entre le Maroc et sa profondeur africaine ».\nPour Alpha Barry, « l’ouverture d’un Consulat général du Burkina Faso vient consolider une relation "déjà très forte" sur le plan économique, comme en témoigne la présence marocaine au Burkina Faso, dans plusieurs domaines tels que les télécommunications, l’industrie, les banques et les assurances ».\nEn matière d’investissement, il ressort qu’« environ 10 % des investissements extérieurs du Royaume dans le continent africain sont concentrés au Burkina Faso, à travers une forte présence des banques et des sociétés marocaines œuvrant dans les domaines des télécommunications, des assurances et de la construction ».\n\nLa cérémonie d’inauguration du Consulat Général du Burkina Faso à Dakhla par les ministres Bourita et Barry\nVisa supprimé\nL’occasion a également été marquée à travers « un accord entre le gouvernement burkinabè et le Royaume du Maroc, portant sur la suppression de visa, en faveur des ressortissants des deux pays titulaires de passeports ordinaires ». Un acte fort apprécié par le ministre Barry. Le Maroc étant l’une des destinations prisées par les Burkinabè.\nSans oublier également que le Maroc octroie annuellement, à des étudiants burkinabè, 150 bourses de formation dans les instituts du Royaume. Dans le cadre de l’inauguration du Consulat à Dakhla, il est convenu « d’octroyer 30 bourses supplémentaires au profit du Burkina Faso dans les instituts de formation dans cette ville, dans des domaines comme le tourisme, l’artisanat, la santé, la gestion et le commerce ».\nLefaso.net
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      Depuis 2016, l’Union européenne (UE) s’est engagée à soutenir les actions de lutte contre le terrorisme menées par le Burkina Faso et les pays du G5 Sahel. Un dialogue continu et une coopération étroite avec les autorités burkinabè, a alors permis d’aboutir à des avancées concrètes, notamment en ce qui concerne le renforcement de capacités des forces de défense et de sécurité face à la menace terroriste. "Malgré une jeune expérience de l’UE dans le domaine de la sécurité, elle a pu apporter son soutien aux forces spéciales à travers la formation et la dotation en équipements, le soutien des FDS à l’est avec le renforcement des casernes et le développement des télécommunications", a affirmé l’ambassadeur Wolfram VETTER.\n\nLe chef de la délégation de l’Union européenne, Wolfram Vetter \nL’UE travaille également à mieux coordonner et rendre plus efficaces ses interventions dans le secteur de la sécurité. C’est dans ce sens que le cadre sectoriel de dialogue "Défense et Sécurité", dont l’institution est actuellement le chef de file pour le compte des partenaires techniques et financiers au Burkina Faso, constitue un cadre privilégié de partage d’informations et de coordination entre les bailleurs. Des initiatives récentes comme le Partenariat pour la Sécurité et la Stabilité au Sahel, lancé par le G7 avec les pays du G5 sahel, visent aussi à rassembler et mettre en cohérence toutes les actions dans la région et au Burkina Faso.\n\n Le ministre de la Sécurité, Ousseni Compaoré \nPour le ministre de la Sécurité, Ousseni Compaoré, l’apport de l’UE dans la lutte contre le terrorisme est énorme et tient compte des besoins exprimés par le Burkina Faso. La présente réunion qui permettra de faire le point des actions menées, sera aussi l’occasion d’explorer de nouvelles perspectives en termes d’actions à initier pour venir à bout du terrorisme.\nArmelle Ouédraogo \nLefaso.net
#> 4                                                                                                                                                                                                                                                                                                                                                                                Cet atelier fait suite à une série de rencontres organisées au profit des communautés vivant au Burkina Faso, par le Ministère de l’Intégration africaine et des Burkinabè de l’extérieur (MIABE), depuis sa création. Au cours de ces différentes rencontres qui ont eu lieu dans plusieurs villes du Burkina Faso, des préoccupations ont été soulevées par les participants. « Les principales préoccupations ont porté sur l’organisation des structures associatives des communautés vivant au Burkina Faso, pour une participation significative au développement socio-économique du pays et particulièrement sur les tracasseries routières ; la non réciprocité d’exemption de visa avec certains pays d’Afrique centrale et son coût élevé ; la non reconnaissance de la carte consulaire par les banques au Burkina Faso ; l’existence de réseaux de trafic de mineurs à des fins de prostitution organisés du Nigeria vers le Burkina Faso », a souligné le ministre Paul Robert Tiendrébéogo.\nC’est pour ainsi répondre à toutes ces préoccupations que le ministère a décidé d’organiser ces ateliers d’information et de sensibilisation, sur les droits et devoirs des communautés vivant au Burkina Faso. Ces ateliers ont eu lieu respectivement le 24 juillet 2020 à Ouagadougou, les 22 et 23 octobre à Banfora et Bobo-Dioulasso.\nLe ministre Tiendrébéogo a expliqué que ces ateliers avaient pour objectifs, d’impliquer les communautés vivant au Burkina Faso, dans la lutte contre le terrorisme et dans la consolidation de la cohésion sociale. Ils visaient également à les sensibiliser sur leurs droits et devoirs dans la consolidation de la démocratie.\n\nDes communautés vivants au Burkina Faso échangent avec le ministre en charge de l’intégration africaine\nMais aussi, à les informer sur les conditions d’accès aux crédits des banques et institutions financières au Burkina Faso, sur les droits d’entrée et d’établissement et les conditions d’applications et d’exemption de visa d’entrée au Burkina Faso.\nCes rencontres déchanges ont enregistré plus de 150 participants, composés de représentants des 16 communautés enregistrées dans les fichiers du ministère en charge de l’Intégration africaine. Il s’agit du Bénin, du Burundi, du Ghana, du Cameroun, de la Côte d’Ivoire, du Mali, du Niger, dud Nigéria, de la République de Guinée, la République du Congo, la République Centrafricaine, la République Démocratique du Congo, du Sénégal, du Tchad, du Togo et du Rwanda.\nAinsi, trois communications ont été dispensées au cours des travaux à savoir : l’implication des communautés vivant au Burkina Faso dans la lutte contre le terrorisme et la consolidation de la paix et de la cohésion sociale ; les droits et devoirs des communautés vivant au Burkina Faso dans la consolidation de la démocratie ; et les conditions d’entrée, de séjour et d’établissement et les conditions d’applications et d’exemption de visa d’entrée au Burkina Faso.\n\nLe ministre de l’intégration africaine, Paul Robert Tiendrébéogo\nLa rencontre du 24 octobre 2020 avec le ministre Tiendrébéogo a permis de restituer les conclusions issues de ces travaux. Selon le rapport de synthèse présenté, les échanges, au cours de ces ateliers d’information et de sensibilisation, ont porté sur plusieurs points, dont la procédure d’obtention de la nationalité burkinabè, les conditions d’obtention de documents de séjour au Burkina Faso, la question relative à la plateforme d’information au niveau du ministère en charge de l’Intégration africaine, la reconnaissance des cartes consulaires par les agents de contrôle et les institutions financières, etc.\nLes participants n’ont pas manqué de faire des recommandations portant, entre autres, sur la mise en place d’une plateforme d’information au profit des communautés vivant au Burkina Faso au sein du ministère, aussi de permettre aux ressortissants de la CEDEAO, d’acquérir la carte biométrique de la CEDEAO dans les pays de résidence. Ils ont aussi formulé une résolution sur la défense des valeurs de paix, de cohésion sociale et de collaboration avec les forces de défense et de sécurité, en leur apportant le soutien des communautés vivant au Burkina Faso.\n\nLa photo de famille des communautés vivant au Burkina Faso avec leur ministre de tutelle\nCette rencontre a été également une occasion pour ces responsables de communautés, d’avoir des échanges directs avec le ministre Paul Robert Tiendrébéogo, en vue d’apporter des réponses précises à leurs préoccupations. Ce dernier les a rassurés que leurs doléances seront examinées afin d’y apporter des solutions. Par ailleurs, il les a invités à mieux s’organiser au sein de leurs structures associatives afin de contribuer au développement du Burkina Faso. « Nous avons pris notes des recommandations issues de la synthèse des travaux. Elles sont très pertinentes et je prends l’engagement qu’elles seront examinées avec attention. Mais des défis restent à relever dans le domaine de l’organisation et du fonctionnement de leurs structures », a-t-il laissé entendre. Avant d’exprimer sa satisfaction car, dit-il, les objectifs visés par ces ateliers ont été atteints.\nRomuald Dofini\nLefaso.net
#> 5 Cette rencontre vise, entre autres, à outiller les participants sur leur rôle et leur place dans la lutte contre le terrorisme et dans la consolidation de la cohésion sociale au Burkina Faso, sur leurs droits et devoirs dans la consolidation de la démocratie et sur les conditions d’entrée, de séjour et établissement au Burkina Faso, ainsi que les conditions d’applications et d’exemption de visas d’entrée au pays des hommes intègres. Elle sera également une occasion pour ces responsables de communautés, d’avoir des échanges directs avec les structures techniques compétentes identifiées, en vue d’apporter des réponses précises à leurs préoccupations.\n\nLe présidium de la cérémonie d’ouverture des travaux de l’atelier\nEn effet, selon le gouverneur de la région des Cascades, Joséphine Kouara Apiou/Kaboré, les défis de la paix, de la cohésion sociale et de la sécurité ont toujours été au centre des préoccupations du gouvernement burkinabè, en vue d’accélérer le processus d’intégration des peuples en Afrique. « C’est la raison pour laquelle, le Ministère de l’Intégration africaine et des Burkinabè de l’extérieur a, depuis sa création, initié des rencontres périodiques avec les différentes communautés vivant au Burkina Faso », a-t-elle souligné.\n\nJonas Sawadogo, chargé de mission au MIABE communicateur\nA l’en croire, ledit ministère a eu, au cours de l’année 2019, plusieurs rencontres de ce genre. Les principales préoccupations abordées au cours de ces rencontres ont porté sur l’organisation des structures associatives des communautés vivant au Burkina Faso, pour une participation significative au développement socio-économique du pays et particulièrement sur les tracasseries routières ; la non réciprocité d’exemption de visa avec certains pays d’Afrique centrale et son coût élevé ; la non reconnaissance de la carte consulaire par les banques au Burkina Faso ; l’existence de réseaux de trafic de mineurs à des fins de prostitution organisés du Nigeria vers le Burkina Faso, etc.\n\nDes responsables des communautés vivant au Burkina Faso prennent part à l’atelier\nC’est pour ainsi répondre à toutes ces préoccupations que le ministère en charge de l’intégration africaine a décidé d’organiser cet atelier d’information et de sensibilisation sur les droits et devoirs des communautés vivant au Burkina Faso.\nSelon la directrice générale de l’intégration africaine, Bernadette Nikiéma/Tapsoba, cet atelier rentre dans le cadre des activités de promotion de l’intégration du MIABE. « Il s’agit pour nous de sensibiliser les communautés vivant au Burkina Faso sur tout ce qui concerne la vie politique, administrative dans notre pays. Nous allons échanger avec ces communautés, sur un certain nombre de préoccupations qu’elles avaient évoquées afin d’y apporter des solutions. A l’issu de cet atelier, nous voulons ainsi qu’elles connaissent d’avantage leurs droits et devoirs et les documents administratifs et consulaires qu’elles peuvent présenter au niveau des établissements financiers, au niveau des autorités administratives du Burkina Faso, pour l’obtention de certains documents et de certaines ressources financières. Nous voulons qu’elles sachent que le Burkina les considère à part entière comme des Burkinabè, et veut avancer dans le développement avec elles », a laissé entendre Bernadette Nikiéma/Tapsoba.\n\nLa directrice générale de l’intégration africaine, Bernadette Nikiéma/Tapsoba\nDes communications spécifiques\nTrois communications seront dispensées au cours des travaux à savoir : l’implication des communautés vivant au Burkina Faso dans la lutte contre le terrorisme et la consolidation de la paix et de la cohésion sociale ; les droits et devoirs des communautés vivant au Burkina Faso dans la consolidation de la démocratie ; et les conditions d’entrée, de séjour et d’établissement et les conditions d’applications et d’exemption de visa d’entrée au Burkina Faso.\nLa première communication est présentée par Sibdou Victorine Ouédraogo, directrice générale de la promotion du dialogue social et de la lutte contre l’exclusion au ministère en charge de l’administration territoriale. Il s’agit pour elle, de montrer aux participants, leur rôle dans le sens de la promotion de la cohésion sociale et dans la lutte contre le terrorisme. « Nous allons d’abord identifier les valeurs communes de vivre-ensemble au Burkina Faso. Ces valeurs consistent à cultiver la solidarité, la tolérance, l’entraide et l’amour au travail. Nous allons leur enseigner ces bonnes valeurs afin que ces communautés ne soient pas des acteurs de désunion pour favoriser l’encrage du terrorisme au Burkina Faso.\n\nSibdou Victorine Ouédraogo, du ministère en charge de l’administration territoriale\nQuant à la deuxième communication, elle est présentée par Jonas Sawadogo, chargé de mission au MIABE. Il a eu la mission d’outiller les participants de l’atelier, sur leurs droits et devoirs et plus particulièrement sur les droits électoraux. C’est-à-dire, leur expliquer le processus de leur participation aux élections de notre pays. « Nous allons leur expliquer comment est-ce qu’ils (membres de ces communautés) peuvent jouir de leurs droits dans notre pays, parce que ce sont ces droits politiques qui les permettent de participer aux élections. Ensuite nous allons leur montrer les étapes possibles pour pouvoir être au niveau du fichier électorale et participer à l’ensemble des scrutins », a dit Jonas Sawadogo.[ Cliquez ici pour lire l’intégralité ]\nRomuald Dofini\nLefaso.net
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         La Ministre du développement de l’économie numérique et des postes Hadja Fatoumata Ouattara Sanon a procédé au lancement officiel du projet « Wuli » de l’incubateur Sira Labs le lundi 26 octobre 2020 à Bobo-Dioulasso. Ce projet financé par la Banque mondiale va outiller gratuitement des milliers de personnes sur les métiers et autres avantages numériques pour le développement du Burkina. \nL’objectif principal du projet Wuli, composante du projet E-Burkina initié par l’Agence nationale de la promotion des TICs (ANPTIC), est d’identifier des idées innovantes et d’accompagner les promoteurs pour en faire des entreprises pérennes dans le domaine du numérique. Précisément dans les secteurs clés de l’économie comme l’agriculture, l’environnement, la santé et l’éducation à Bobo-Dioulasso.\n\n\n(adsbygoogle = window.adsbygoogle || []).push({});\n\nSe référant aux règles élémentaires de l’économie, la Ministre du développement de l’économie numérique et des postes Hadja Fatoumata Sanon a rappelé la contribution significative que le projet « Wuli » va apporter dans  la recherche de la souveraineté numérique de notre pays.\nLes participants au lancement« Il faut qu’on commence à consommer ce que nous produisons. Mais cela suppose qu’on doit d’abord œuvrer de sorte à ce qu’il y ait une masse critique de Burkinabè qui construisent des plateformes  dont nous avons besoin»,  explique-t-elle. A l’en croire, cette initiative qui entre dans le cadre du projet E-Burkina  va faire la part belle au monde rural et agricole. \nUn ciblage qui, dira-t-elle, correspond bien aux réalités du pays où l’agriculture occupe plus de 80% de la population. Il sera aussi question de lutter contre la pauvreté féminine à travers la fixation d’un quota de 30% de filles.\nUn processus qui va évoluer\nPendant deux ans, les bénéficiaires du projet « Wuli » auront un cadre technologique approprié et du suivi de coachs à leurs dispositions. Chose qui permettra de booster leurs innovations. Selon Hadja Fatoumata Ouattara Sanon, ce projet de vulgarisation des TICs et de développement des startups numériques va évoluer de la création des E-services à la construction d’appareils informatiques.\n\nC’est gratuit\n« Les bénéficiaires s’inscrivent gratuitement sur le site www.sira-labs.com. On les stimule, on les accompagne à mûrir leurs projet et à développer leurs entreprises numériques à travers trois volets dont la formation, le suivi-conseil et le financement qui sera matérialisé par un soutien cash ou la dotation en infrastructure. A la fin, nous aurons au moins 15 startups numériques et 3000 jeunes formés. Que ce soit les fonctionnaires, les étudiants, les agriculteurs, les commerçants, nous allons inclure tout le monde dans ce processus de développement numérique »\nLas Django Bathily, DG de Sira \n\nWuli ka sira ta (Se lever pour entreprendre en langue Bambana ou Dioula), qui a donné son nom au projet « Wuli », est exécuté par l’incubateur Sira Labs. Pour son  Directeur général  Django Bathily, l’impact de l’entrepreneuriat numérique va au-delà des emplois virtuels. «Parce que le numérique est transversal » dans la mesure où la création d’un projet numérique crée des emplois dans d’autres secteurs.\nEstimé à un coût total de 275 millions de francs CFA, le projet « Wuli » a été financé par la Banque mondiale dont la Représentante résidente au Burkina Faso Maimouna Mbow Fam a participé à la cérémonie de lancement à travers un système de vidéo-conférence.  « L’incubateur de Bobo va couvrir toute la partie ouest du Burkina en attendant l’extension du projet aux autres régions. Car en plus des infrastructures et autres projets structurants, la Banque mondiale soutient également le développement des projets innovants », rassure Maimouna Mbow Fam.\nAminata SANOU\nCorrespondante de Burkina 24 à Bobo-Dioulasso
#>   press_article
#> 1       fasonet
#> 2       fasonet
#> 3       fasonet
#> 4       fasonet
#> 5       fasonet
#> 6     burkina24