Main functions of collecting news from Côte Ivoir

Function for collecting news (title + content) from the media in Côte Ivoir.

ci_news_contents(start_d, end_d, press = "all")

Arguments

start_d
yyyy-mm-dd start date of collection (oldest date, eg: yesterday)

end_d
yyyy-mm-dd collection end date (newest date, eg: today)

press
specify the name of a press.

Value

a table containing the news (title + content + date publisher + press) collected

Examples

library(afnews)
dt_news <- ci_news_contents(start_d="2020-10-20", end_d="2020-10-25",press='all')

#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> Done!
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> Done!
#> Done!

head(dt_news)

#>   date_article
#> 1   2020-10-25
#> 2   2020-10-22
#> 3   2020-10-21
#> 4   2020-10-20
#> 5   2020-10-21
#> 6   2020-10-20
#>                                                                                                                                                                                           url_article
#> 1                                                                 https://www.linfodrome.com/afrique-monde/62293-cameroun-les-hommes-en-armes-attaquent-une-ecole-primaire-plusieurs-morts-et-blesses
#> 2 https://www.linfodrome.com/revue-de-presse-nationale/62182-presidentielle-2020-dabou-assiege-au-moment-ou-l-opposition-demande-d-intensifier-la-desobeissance-civile-le-gouvernement-veut-dialoguer
#> 3                                                    https://www.linfodrome.com/revue-de-presse-nationale/62139-des-loubards-neutralises-sur-le-campus-grosse-colere-de-l-opposition-contre-la-cedeao
#> 4                                                https://www.linfodrome.com/revue-de-presse-nationale/62112-cote-d-ivoire-tout-le-pays-s-embrase-la-cedeao-gifle-bedie-et-affi-la-fesci-paralyse-tout
#> 5                                                                        https://aip.ci/cote-divoire-le-conseil-cafe-cacao-offre-2000-kits-scolaires-et-des-tables-bancs-aux-enfants-des-producteurs/
#> 6                                                                                                                  https://aip.ci/cote-divoire-aip-un-projet-de-greffage-du-karite-lance-a-boundiali/
#>                                                                                                   title_article
#> 1                        Cameroun : Les hommes en armes attaquent une école primaire plusieurs morts et blessés
#> 2                 Présidentielle 2020 : Dabou assiégé au moment où l’opposition demande d’intensifier la dés...
#> 3                       Des loubards neutralisés sur le campus ; Grosse colère de l’opposition contre la Cedeao
#> 4                 Côte d’Ivoire : Tout le pays s’embrase, la Cedeao ‘’gifle Bédié et Affi, La Fesci paralyse...
#> 5 Côte d’Ivoire/Le Conseil café-cacao offre 2000 kits scolaires et des tables-bancs aux enfants des producteurs
#> 6                                          Côte d’Ivoire-AIP/ Un projet de greffage du karité lancé à Boundiali
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      Content_article
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               Le premier bilan officiel de l’attaque d’une école primaire à Kumba (Sud-ouest) samedi fait état de 6 morts et 13 blessés a déclaré le gouvernement camerounais.if (!window.AdButler){(function(){var s =document.createElement("script"); s.async = true; s.type ="text/javascript";s.src = 'https://servedbyadbutler.com/app.js';var n =document.getElementsByTagName("script")[0]; n.parentNode.insertBefore(s,n);}());}var AdButler = AdButler || {}; AdButler.ads = AdButler.ads || [];var abkw = window.abkw || '';var plc440701 = window.plc440701 || 0;document.write('<'+'div id="placement_440701_'+plc440701+'"></'+'div>');AdButler.ads.push({handler: function(opt){ AdButler.register(178368, 440701,[300,250], 'placement_440701_'+opt.place, opt); }, opt: { place:plc440701++, keywords: abkw, domain: 'servedbyadbutler.com',click:'CLICK_MACRO_PLACEHOLDER' }});Selon le gouvernement qui a dressé tard dans la nuit de samedi à dimanche, le premier bilan de l’attaque dans un établissement scolaire situé au quartier Fiango, arrondissement de Kumba 2 (Sud-Ouest), 6 élèves bien identifiés dont 1 garçon et 5 filles ont été tués, tous âgés de neuf à douze ans.René Emmanuel Sadi le ministre de la communication et porte-parole du gouvernement révèle également que 13 élèves ont été blessés soit 10 filles et 03 garçons, dont sept cas avérés préoccupants.Selon le gouvernement, l’attaque est l’œuvre d’un groupe de près d’une dizaine de terroristes, munis d’armes de guerre et constitués en véritable commando. Ledit groupe a fait irruption aux environs de 11 locales (10hTU), à bord de trois motocyclettes, dans l’enceinte du complexe scolaire privé dénommé « Mother Francisa International Bilingual Academy ».Par ailleurs, ledit établissement scolaire n’a lancé ses activités qu’en ce début d’année scolaire 2020/2021, « à l’insu des autorités administratives compétentes, et n’a pu bénéficier des mêmes mesures de protection que d’autres établissements scolaires du département de la Mémé », relève Yaoundé. Le gouvernement camerounais estime qu’aucune « cause au monde ne peut justifier ou légitimer de telles atteintes systématiques à la vie d’êtres humains, encore moins, à celle d’enfants innocents, d’élèves, ou d’étudiants, qui doivent aller à l’école, au collège, au lycée, à l’université, pour se former et être en mesure de se prendre en charge, et d’assumer dans le futur la relève de leurs aînés.» Le gouvernement camerounais a également appelé à la coopération internationale et à l’aide des pays amis qui hébergent ceux qui soutiennent, alimentent, financent le terrorisme au Cameroun et le climat d’insécurité dans les deux régions du Nord-ouest et du Sud-Ouest. Selon les autorités camerounaises, il ne fait pas de doute que l’attentat terroriste commis dans la ville de Kumba, traduit le « désarroi et la déroute des bandes armées sécessionnistes face à la détermination de la communauté éducative, de braver leurs menaces et leurs mots d’ordre, en décidant la reprise des classes dans leurs localités respectives.» Yaoundé pense que l’objectif des sécessionnistes est de briser la dynamique observée dans la reprise des classes dans le cadre de la rentrée scolaire 2020/2021 dans les régions du Nord-Ouest et du Sud-Ouest. Condamnations L’ensemble de la classe politique camerounaise a condamné un acte « terroriste d’une cruauté et d’une barbarie insoutenables » contre des jeunes innocents.Parfait KOFFI (function(d, s, id) {\n\t\t\t\t  var js, fjs = d.getElementsByTagName(s)[0];\n\t\t\t\t  if (d.getElementById(id)) {return;}\n\t\t\t\t  js = d.createElement(s); js.id = id;\n\t\t\t\t  js.src = "//connect.facebook.net/fr_FR/all.js#xfbml=1&appId=607728835948702";\n\t\t\t\t  fjs.parentNode.insertBefore(js, fjs);\n\t\t\t\t}(document, 'script', 'facebook-jssdk'));
#> 2 Les médias ivoiriens ont consacré ce jeudi 22 octobre 2020 leurs publications aux affrontements à Dabou, à la sortie de Pascal Affi N’guessan appelant à intensifier la désobéissance civile et au dialogue entre le gouvernement et l’opposition.if (!window.AdButler){(function(){var s =document.createElement("script"); s.async = true; s.type ="text/javascript";s.src = 'https://servedbyadbutler.com/app.js';var n =document.getElementsByTagName("script")[0]; n.parentNode.insertBefore(s,n);}());}var AdButler = AdButler || {}; AdButler.ads = AdButler.ads || [];var abkw = window.abkw || '';var plc440701 = window.plc440701 || 0;document.write('<'+'div id="placement_440701_'+plc440701+'"></'+'div>');AdButler.ads.push({handler: function(opt){ AdButler.register(178368, 440701,[300,250], 'placement_440701_'+opt.place, opt); }, opt: { place:plc440701++, keywords: abkw, domain: 'servedbyadbutler.com',click:'CLICK_MACRO_PLACEHOLDER' }});A quelques jours de la présidentielle 2020, alors que les candidats Alassane Ouattara et Kouadio Konan Bertin dit KKB sont en pleine campagne, certaines régions du pays sont en ébullition, suite à l’appel à la désobéissance civile lancée par les plateformes de l’opposition ivoirienne. Après Bongouanou, les villes de Bonoua et de Dabou sont devenues le théâtre des affrontements entre autochtones et allochtones. La région du Leboutou était particulièrement à feu et à sang, le mercredi 21 octobre 2020, et il a fallu y déployer des chars pour ramener un tant soit peu le calme. « Dabou à feu et à sang, hier : Des chars contre les populations aux mains nues », titre Le Quotidien d’Abidjan, journal proche de l’opposition qui annonce « Plusieurs morts, des blessés graves et des biens détruits ». Pour Le Temps, autre quotidien proche de l’opposition, « Dabou » est une « ville assiégée ». Le confrère nous apprend que « La ville de Dabou, selon des informations en provenance de la capitale du Leboutou, est assiégée depuis hier mercredi 21 octobre 2020, par les forces de défense et de sécurité. Cette présence massive des forces est d’éviter un affrontement ». Pour sa part, le confrère L’inter, quotidien indépendant nous annonce l’instauration d’un « Couvre-feu à Dabou » « à quelques jours de la présidentielle ». Selon le confrère, « Des jeunes en majorité, vêtus pour certains en rouge [symbole du guerrier] ont constitué des groupes pour, disent-ils, ‘’venir se défendre contre des agresseurs’’ qui auraient pris en otage leur ville : ‘’Nous sommes venus marcher, eux, ils sont venus nous agresser’’, déclarait un jeune manifestant surexcité, dans une vidéo, tandis que les forces de l’ordre tentaient de calmer les ardeurs ». Rapportant les propos d’Adama Bictogo, directeur national de campagne RHDP, L’Expression nous informe que « Dabou, c’est du terrorisme…L’Etat va sévir ». Selon le confrère, Bictogo a fait cette déclaration, le mercredi 21 octobre 2020, au siège du RHDP des II Plateaux Vallons où il a fait le point de la campagne. « Ce qui se passe à Dabou est la conséquence de l’appel de Soro à travers Ben Souk. L’opposition recherche un affrontement communautaire. L’Etat va sévir ».La presse ivoirienne s’est également appesantie sur la conférence de presse du président du FPI, Pascal Affi N’guessan, tenue le mercredi 21 octobre 2020, au siège du PDCI-RDA. Sur ce sujet, Le Temps barre à sa Une « Manifestations contre le 3ème mandat/ Affi : ‘’Abidjan va bouger, le fruit va tomber’’… » et rapporte que « Il faut intensifier la désobéissance civile ». Dans la même dynamique, Le Quotidien d’Abidjan titre. « Conférence de presse de l’opposition, hier/ Affi : ‘’Quand Abidjan va bouger, le fruit va tomber’’ ». Toujours selon ce confrère qui rapporte les propos de Pascal Affi N’guessan, « Il y aura forcément une transition », « Il n’y aura pas d’élection le 31 octobre », « Nous ne participerons pas à une rencontre avec le pouvoir sans l’ONU ».Le Nouveau Réveil, quotidien proche du PDCI-RDA abonde dans le même sens, en rapportant les propos d’Affi N’guessan qui a parlé « au nom de toute l’opposition face à la presse, hier : ‘’Le 31 octobre, nous passons à la transition, si… ». Le 3ème sujet qui revient dans la presse ivoirienne est le dialogue entre le pouvoir et l’opposition. « Un important message du gouvernement est tombé, hier », selon le quotidien indépendant Soir Info qui ajoute que « Ouattara (est) prêt à rencontrer Bédié ». Le confrère nous apprend que « Dans un communiqué publié, le mercredi 21 octobre 2020, le gouvernement a annoncé sa disponibilité à accéder aux requêtes de l’opposition sur la recomposition de la commission électorale indépendante (CEI) centrale et des CEI locales »A propos « Du dialogue politique avec l’opposition », Le Patriote, quotidien proche du RHDP, nous apprend que « Le gouvernement tend la main…Et fait des concessions ». Même son de cloche pour L’Expression. « Dialogue politique, hier à la Primature : Le PDCI va reprendre sa place à la CEI ». Adolphe AngouaSauf autorisation de la rédaction ou partenariat pré-établi, la reprise des articles de linfodrome.com, même partielle, est strictement interdite. Tout contrevenant s’expose à des poursuites. (function(d, s, id) {\n\t\t\t\t  var js, fjs = d.getElementsByTagName(s)[0];\n\t\t\t\t  if (d.getElementById(id)) {return;}\n\t\t\t\t  js = d.createElement(s); js.id = id;\n\t\t\t\t  js.src = "//connect.facebook.net/fr_FR/all.js#xfbml=1&appId=607728835948702";\n\t\t\t\t  fjs.parentNode.insertBefore(js, fjs);\n\t\t\t\t}(document, 'script', 'facebook-jssdk'));
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                La presse ivoirienne rapporte ce mercredi 21 octobre 2020 que des loudards ont été neutralisés sur le campus et la grosse colère de l’opposition contre la Cedeao.if (!window.AdButler){(function(){var s =document.createElement("script"); s.async = true; s.type ="text/javascript";s.src = 'https://servedbyadbutler.com/app.js';var n =document.getElementsByTagName("script")[0]; n.parentNode.insertBefore(s,n);}());}var AdButler = AdButler || {}; AdButler.ads = AdButler.ads || [];var abkw = window.abkw || '';var plc440701 = window.plc440701 || 0;document.write('<'+'div id="placement_440701_'+plc440701+'"></'+'div>');AdButler.ads.push({handler: function(opt){ AdButler.register(178368, 440701,[300,250], 'placement_440701_'+opt.place, opt); }, opt: { place:plc440701++, keywords: abkw, domain: 'servedbyadbutler.com',click:'CLICK_MACRO_PLACEHOLDER' }});« Attaque de l’Université de Cocody : Des loudards et microbes neutralisés sur le campus », titre Le Temps. Qui informe que des loubards et microbes ont fait irruption à l’Université de Cocody, le mardi 20 octobre 2020. Pour s’attaquer aux membres de la Fédération estudiantine et scolaire de Côte d’Ivoire (Fesci), qui avait lancé un mot d’ordre de grève le lundi dans tous les établissements publics et privés. Pour protester contre les frais annexes d’inscription et les cotisations Coges. « Inscrits dans une logique de combattre tous ceux qui font ombrage à leur régime, des loubards et microbes payés par des pontes du régime, se sont violemment attaqués aux étudiants au campus de Cocody. Mal leur en prit, car les membres de la Fesci, en territoire, ont organisé la riposte », souligne Le Temps.Pour sa part, Linter note la « grosse colère de l’opposition contre la Cedeao ». Le confrère rapporte la réaction de l’opposition après la dernière mission de la Cedeao à Abidjan. « Mardi 20 octobre 2020, en conférence de presse, à Abidjan, le camp de l’ancien président Laurent Gabgbo a laissé éclater sa colère devant ce qu’il conçoit comme une honte pour le continent africain », fait savoir ce journal.Quant à Fraternité Matin, il barre à sa Une : « Accès au financement, plan de relance post-covid-19 : Patrick Achi présente les ambitions du candidat Ouattara au Patronat ». Ce quotidien pro-gouvernemental apprend que le Directeur national de campagne du porte-flambeau du Rhdp à l’élection présidentielle a entretenu les dirigeants d’entreprise sur les actions du chef de l’État.Terminons cette revue de presse par ce fait tragique livré par Soir Info : « Un employé d’une société de dragage chute dans la lagune et meurt ». « L’entreprise de dragage, installée dernière le camp de gendarmerie commando parachutiste de Koumassi a perdu l’un de ses employés, de manière tragique. Il est décédé à la tâche, pourrait-on dire », rapporte ce journal indépendant de Côte d’Ivoire.Jonas BAIKEH  (function(d, s, id) {\n\t\t\t\t  var js, fjs = d.getElementsByTagName(s)[0];\n\t\t\t\t  if (d.getElementById(id)) {return;}\n\t\t\t\t  js = d.createElement(s); js.id = id;\n\t\t\t\t  js.src = "//connect.facebook.net/fr_FR/all.js#xfbml=1&appId=607728835948702";\n\t\t\t\t  fjs.parentNode.insertBefore(js, fjs);\n\t\t\t\t}(document, 'script', 'facebook-jssdk'));
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 La tension était vive en Côte d’Ivoire dans la journée du lundi 19 octobre 2020. Plusieurs villes bouillaient de violences. Et la Cedeao s’est prononcée sur cet état de fait. Les médias ivoiriens ont relayé les péripéties de cette journée chaude ce mardi 20 octobre 2020.if (!window.AdButler){(function(){var s =document.createElement("script"); s.async = true; s.type ="text/javascript";s.src = 'https://servedbyadbutler.com/app.js';var n =document.getElementsByTagName("script")[0]; n.parentNode.insertBefore(s,n);}());}var AdButler = AdButler || {}; AdButler.ads = AdButler.ads || [];var abkw = window.abkw || '';var plc440701 = window.plc440701 || 0;document.write('<'+'div id="placement_440701_'+plc440701+'"></'+'div>');AdButler.ads.push({handler: function(opt){ AdButler.register(178368, 440701,[300,250], 'placement_440701_'+opt.place, opt); }, opt: { place:plc440701++, keywords: abkw, domain: 'servedbyadbutler.com',click:'CLICK_MACRO_PLACEHOLDER' }});Les manifestations qui ont eu lieu dans tout le pays ont laissé des traces tant psychologiques que physiques. Des dégâts et non des moindres ont été constatés partout où ces heurts se sont déroulés. « 3e mandat anticonstitutionnel : Tout le pays s’embrase » barre Le Quotidien d’Abidjan qui poursuit « Bonoua : 1 mort et des blessés graves », « Abidjan : des véhicules endommagés et la circulation perturbée ». Dans la même veine, Le Temps met également en grand les conséquences des événements de la veille. « Désobéissance civile : plusieurs villes paralysées, hier » parle le quotidien. Pour Le Nouveau Réveil, c’est toute « la Côte d’Ivoire en ébullition, hier » qu’il traduit.Pendant que les manifestations ont cours dans le pays, une délégation de la Cedeao rencontrait le Chef de l’Etat et le Premier ministre pour discuter de la situation sociopolitique qui prévaut en Côte d’Ivoire. au sortir des rencontres, Le patriote a retenu de la déclaration de cette délégation que « La Cedeao ‘’gifle’’ Bédié et Affi ». L’expression met l’accent sur « Bloquer les rues et saisir le matériel électoral sont des crimes », une parole forte de la délégation de « la Cedeao à Bédié et Affi ». Cela sonne comme un désaveu de la Cedeao à l’encontre des deux leaders, fait savoir Le Jour plus. « Présidentielle du 31 octobre : La Cedeao désavoue Affi et Bédié » explique le confrère.La société était encore sous tension à cause de la Fédération estudiantine et scolaire de Côte d’Ivoire (Fesci). Elle est entrée en grève le lundi 19 octobre 2020. Les écoles ont subi les conséquences de leur courroux face à la ministre Camara Kandia. « 1ere journée de grève : La Fesci a paralysé toutes les écoles du pays » affirme Le Nouveau Réveil. Le Quotidien d’Abidjan évoque les raisons de cette paralysie. « Cotisation exagérée dans les écoles la Fesci paralyse tout » écrit le journal qui continue « le mouvement de grève largement suivi ».Alain Zadi (function(d, s, id) {\n\t\t\t\t  var js, fjs = d.getElementsByTagName(s)[0];\n\t\t\t\t  if (d.getElementById(id)) {return;}\n\t\t\t\t  js = d.createElement(s); js.id = id;\n\t\t\t\t  js.src = "//connect.facebook.net/fr_FR/all.js#xfbml=1&appId=607728835948702";\n\t\t\t\t  fjs.parentNode.insertBefore(js, fjs);\n\t\t\t\t}(document, 'script', 'facebook-jssdk'));
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         Quelques récipiendaires des kits et tenues scolaires du conseil café-cacao avec la délégation de M. Koné et les enseignants à Lassinabango San-Pedro, 21 oct 2020 (AIP)- Le Conseil café-cacao offre, pour l’année scolaire 2020-2021, 2000 kits et 355 tenues scolaires ainsi que 350 tables-bancs aux élèves et aux écoles primaires des localités du Bas-Sassandra, pour soutenir les producteurs à la scolarisation de leurs enfants, a fait savoir, mardi 20 octobre 2020, à Lassinabango (village de la sous-préfecture de Gabiadji), le délégué régional de San-Pedro,  Koné Marcel. Rencontré à Lassinabango, dans le cadre d’une tournée de distribution de ces dons, le délégué régional du Conseil café-cacao a indiqué que ce don vise à inciter les producteurs de cacao et du café à scolariser leurs enfants. Les établissements bénéficiaires ont été identifiés par  40 sociétés coopératives partenaires désignées... Sauf autorisation de la rédaction ou partenariat pré-établi, la reprise des articles de AIP.CI même partielle, est strictement interdite. Tout contrevenant s'expose à des poursuites. Veuillez vous connecter pour lire l'intégralité. Nom d’utilisateur ou adresse e-mail  Mot de passe    Se souvenir de moi  Mot de passe oublié? Vous abonner
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             Des productrices de karité venues massivement au lancement du projet. Boundiali, 20 oct 2020 (AIP) – Le doyen de l’Unité de formation régionale des sciences biologiques de l’Université Peleforo Gon Coulibaly (UPGC), Professeur Dirrassouba Nafan, a procédé au lancement du projet intitulé “Transfert de la technique de greffage de karité en milieu paysan dans la région de la Bagoué”. Ce projet s’inscrit dans le cadre du Programme filières agricoles durables de Cote d’Ivoire (FADCI), soutenu par le Fonds compétitif pour l’Innovation agricole durable (FCIAD) et exécuté par le Fonds interprofessionnel pour la recherche et le conseil agricoles (FIRCA). Ce lancement s’est fait mercredi 15 octobre 2020 à la mairie de Boundiali, en présence d’une cinquantaine de productrices de beurre de karité issues des départements de Tengréla,... Sauf autorisation de la rédaction ou partenariat pré-établi, la reprise des articles de AIP.CI même partielle, est strictement interdite. Tout contrevenant s'expose à des poursuites. Veuillez vous connecter pour lire l'intégralité. Nom d’utilisateur ou adresse e-mail  Mot de passe    Se souvenir de moi  Mot de passe oublié? Vous abonner
#>   press_article
#> 1    linfodrome
#> 2    linfodrome
#> 3    linfodrome
#> 4    linfodrome
#> 5           api
#> 6           api