Main functions of collecting information on Senegal

Function for collecting news (title + content) from the media in Burkina Faso.

sn_news_contents(start_d, end_d, press = "all")

Arguments

start_d
yyyy-mm-dd start date of collection (oldest date, eg: yesterday)

end_d
yyyy-mm-dd collection end date (newest date, eg: today)

press
specify the name of a press.

Value

a table containing the news (title + content + date publisher + press) collected

Examples

library(afnews)
dt_news <- sn_news_contents(start_d="2020-10-28", end_d="2020-11-02", press='all')

#> Done!

#> Warning: No article publish in this period !

#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>> 
#> Done!

head(dt_news)

#>   date_article
#> 1   2020-11-02
#> 2   2020-11-02
#> 3   2020-11-02
#> 4   2020-11-02
#> 5   2020-11-02
#> 6   2020-11-01
#>                                                                                                          url_article
#> 1                         http://lesoleil.sn/guinee-lopposition-conteste-devant-la-justice-la-victoire-dalpha-conde/
#> 2           http://lesoleil.sn/retrouvailles-macky-idy-garder-le-pouvoir-pendant-50-ans-selon-serigne-mbacke-ndiaye/
#> 3                         http://lesoleil.sn/guinee-lopposition-conteste-devant-la-justice-la-victoire-dalpha-conde/
#> 4 http://lesoleil.sn/presidentielle-americaine-i-francois-durpaire-historien-trump-na-eu-aucune-politique-africaine/
#> 5                      http://lesoleil.sn/tanzanie-tundu-lissu-candidat-battu-a-la-presidentielle-arrete-a-son-tour/
#> 6                              http://lesoleil.sn/le-senegal-condamne-fermement-les-attentats-terroristes-en-france/
#>                                                                                                  title_article
#> 1                                    Guinée: l’opposition conteste devant la justice la victoire d’Alpha Condé
#> 2                     Retrouvailles Macky-Idy: «garder le pouvoir pendant 50 ans», selon Serigne Mbacké Ndiaye
#> 3                                    Guinée: l’opposition conteste devant la justice la victoire d’Alpha Condé
#> 4 Présidentielle américaine I François Durpaire, Historien: « la gestion de la pandémie peut pénaliser Trump »
#> 5                                 Tanzanie: Tundu Lissu, candidat battu à la présidentielle, arrêté à son tour
#> 6                                        Le Sénégal « condamne fermement » les attentats terroristes en France
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  Content_article
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    Quatre candidats déclarés perdants à l’issue de la présidentielle en Guinée, dont Cellou Dalein Diallo, arrivé deuxième selon l’instance chargée d’organiser les élections, ont saisi la plus haute juridiction du pays pour contester la victoire du président sortant Alpha Condé, a-t-on appris lundi. Cellou Dalein Diallo, qui a obtenu 33,5% des suffrages le 18 octobre, selon les résultats provisoires proclamés par la commission électorale (Céni), contre 59,5% pour M. Diallo, dispose de « preuves irréfutables » que le scrutin a été entaché « d’irrégularités », a déclaré à l’AFP son avocat, Amadou Diallo. Ces preuves, qui n’ont pas été détaillées, doivent selon l’avocat « permettre à la Cour constitutionnelle d’accéder à notre requête », déposée dimanche: « l’annulation du vote dans la région de Kankan », fief de M. Condé dans l’est du pays, et la « proclamation de la victoire dès le premier tour » de Cellou Dalein Diallo, adversaire de longue date d’Alpha Condé. Sans attendre l’annonce de résultats, M. Diallo avait clamé sa victoire dès le 19 octobre, affirmant avoir remporté 53% des suffrages. Des heurts avaient ensuite éclaté, durant une semaine avant de progressivement s’apaiser. Trois autres des douze candidats ont également introduit un recours, selon une source au greffe de la Cour constitutionnelle: Ousmane Kaba (1,19% des voix), Ibrahima Abé Sylla (1,55%) et Makalé Traoré (2,72%). La Cour, en « télétravail » depuis lundi après qu’une de ses juges a contracté le coronavirus, selon la même source, dispose d’un délai de cinq jours pour statuer définitivement sur l’issue d’une élection qui s’est déroulée dans un contexte très tendu, comme c’est souvent le cas dans ce pays d’Afrique de l’Ouest. La France, l’Union européenne et les Etats-Unis ont mis en doute la « crédibilité » du résultat annoncé par la Céni. Cellou Dalein Diallo a pour sa part déjà jugé improbable que la cour invalide le scrutin. Plusieurs candidats de l’opposition ont par ailleurs appelé vendredi à la reprise des manifestations anti-Condé dès mardi pour, selon M. Diallo, « dénoncer le hold-up électoral » et « exiger la reconnaissance » de sa « victoire ». Selon le gouvernement, les violences post-électorales ont fait 21 morts, dont des membres des forces de l’ordre. L’opposition dénonce pour sa part une « répression sanglante » ayant fait « plus de 30 » tués, selon M. Diallo. Réélu en 2015 pour un deuxième et dernier mandat, M. Condé s’est représenté pour un troisième, arguant que l’adoption en mars d’une nouvelle Constitution remettait à zéro le compteur des mandats présidentiels. Mais ses adversaires ont dénoncé un « coup d’Etat constitutionnel ». AFP
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           Serigne Mbacké Ndiaye s’est «réjoui» du rapprochement entre ses «frères» Macky Sall et Idrissa Seck acté, dimanche, à travers l’entrée du parti Rewmi dans le gouvernement. «Macky (Sall) est aujourd’hui le leader de la famille mais Wade en est le père. Macky-Wade-Idy réunis, c’est la poursuite de l’objectif de garder le pouvoir pendant cinquante ans et ceci dans l’intérêt supérieur de la Nation. Je ne tire pas la couverture de mon côté mais je m’en réjouis et j’exprime ma satisfaction», a réagi notamment Serigne Mbacké Ndiaye suite à la nomination du président du Conseil départemental de Thiès à la tête du Conseil économique, social et environnemental (Cese).\nEt M. Ndiaye d’ajouter : « Quelques jours après la perte du pouvoir, en 2012, et au cours d’une réunion du Comité directeur du PDS tenue au domicile du Président Wade à Fann, je défendais l’idée des retrouvailles de la grande famille libérale. Je n’avais pas été compris mais le temps a fait son travail».\nA ce titre, renchérit l’ancien porte-parole du Pape du Sopi, «je suis heureux que ces deux frères se retrouvent. Une partie du chemin vient d’être fait. L’autre partie non moins importante consiste à se retrouver avec et autour du Président Abdoulaye Wade, père du Libéralisme en Afrique Noire et au Sénégal».
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    Quatre candidats déclarés perdants à l’issue de la présidentielle en Guinée, dont Cellou Dalein Diallo, arrivé deuxième selon l’instance chargée d’organiser les élections, ont saisi la plus haute juridiction du pays pour contester la victoire du président sortant Alpha Condé, a-t-on appris lundi. Cellou Dalein Diallo, qui a obtenu 33,5% des suffrages le 18 octobre, selon les résultats provisoires proclamés par la commission électorale (Céni), contre 59,5% pour M. Diallo, dispose de « preuves irréfutables » que le scrutin a été entaché « d’irrégularités », a déclaré à l’AFP son avocat, Amadou Diallo. Ces preuves, qui n’ont pas été détaillées, doivent selon l’avocat « permettre à la Cour constitutionnelle d’accéder à notre requête », déposée dimanche: « l’annulation du vote dans la région de Kankan », fief de M. Condé dans l’est du pays, et la « proclamation de la victoire dès le premier tour » de Cellou Dalein Diallo, adversaire de longue date d’Alpha Condé. Sans attendre l’annonce de résultats, M. Diallo avait clamé sa victoire dès le 19 octobre, affirmant avoir remporté 53% des suffrages. Des heurts avaient ensuite éclaté, durant une semaine avant de progressivement s’apaiser. Trois autres des douze candidats ont également introduit un recours, selon une source au greffe de la Cour constitutionnelle: Ousmane Kaba (1,19% des voix), Ibrahima Abé Sylla (1,55%) et Makalé Traoré (2,72%). La Cour, en « télétravail » depuis lundi après qu’une de ses juges a contracté le coronavirus, selon la même source, dispose d’un délai de cinq jours pour statuer définitivement sur l’issue d’une élection qui s’est déroulée dans un contexte très tendu, comme c’est souvent le cas dans ce pays d’Afrique de l’Ouest. La France, l’Union européenne et les Etats-Unis ont mis en doute la « crédibilité » du résultat annoncé par la Céni. Cellou Dalein Diallo a pour sa part déjà jugé improbable que la cour invalide le scrutin. Plusieurs candidats de l’opposition ont par ailleurs appelé vendredi à la reprise des manifestations anti-Condé dès mardi pour, selon M. Diallo, « dénoncer le hold-up électoral » et « exiger la reconnaissance » de sa « victoire ». Selon le gouvernement, les violences post-électorales ont fait 21 morts, dont des membres des forces de l’ordre. L’opposition dénonce pour sa part une « répression sanglante » ayant fait « plus de 30 » tués, selon M. Diallo. Réélu en 2015 pour un deuxième et dernier mandat, M. Condé s’est représenté pour un troisième, arguant que l’adoption en mars d’une nouvelle Constitution remettait à zéro le compteur des mandats présidentiels. Mais ses adversaires ont dénoncé un « coup d’Etat constitutionnel ». AFP
#> 4 Docteur et agrégé d’Histoire, François Durpaire est Professeur à l’Université de Cergy (France) et auteur de plus d’une quinzaine d’ouvrages. Titulaire d’une thèse sur les relations États-Unis/Afrique, il analyse les enjeux d’une présidentielle américaine particulière. Propos recueillis par Moussa DIOP Le duel Trump-Biden s’inscrit-il dans la continuité des présidentielles américaines ou est-ce une rupture ? C’est une rupture assurément. Jamais dans l’histoire politique des États-Unis, une épidémie n’avait transformé à ce point les modalités du vote. On estime que près des 2/3 des Américains ont déjà voté par anticipation au moment où commence le « jour électoral » (electoral day), le 3 novembre. D’autre part, les votes par correspondance (mail-in ballots) laissent planer un doute sur la date des résultats, car certains États acceptant les enveloppes jusqu’à 10 jours après le scrutin ! La Covid-19 a aussi changé la campagne : pas de « porte-à-porte », des débats avec un plexiglas pour séparer les candidats, peu de vrais meetings du côté démocrate, un Président candidat lui-même hospitalisé à 30 jours du scrutin ! La gestion de la pandémie de coronavirus peut-elle faire pencher l’élection présidentielle ? C’est possible, et de deux manières. Premièrement, la gestion de la pandémie peut pénaliser Trump, le candidat sortant, accusé d’avoir sous-estimé la crise sanitaire et de ne pas prendre les bonnes décisions. Elle a, en outre, détruit une partie de l’économie américaine, et donc, saboté son bon bilan économique. Deuxièmement, la pandémie peut aussi avoir une influence en termes de mobilisation de l’électorat et de contestation éventuelle des résultats, si certains bulletins par correspondance sont jugés défectueux (absence de signature, etc.). À la Maison-Blanche, Trump a survécu à quatre ans de polémiques. N’aurait-il pas une base assez solide de soutiens pour une réélection ? Il y a trois groupes qui soutiennent Trump. Il serait caricatural de considérer que ses soutiens sont tous des suprémacistes blancs sans diplôme. Il y a d’abord des gens satisfaits par son travail économique et notamment par sa politique de baisse d’impôts. La récente hausse du Pib de 33 % (de juillet à septembre) a donné des arguments au Président pour dire que l’économie repartait à la hausse. Il y a ensuite les conservateurs qui se félicitent du discours d’ordre du Président (law and order) et qui sont satisfaits de la nomination de juges conservateurs à la Cour suprême, laissant présager des décisions de restriction du droit à l’avortement. Et il y a, enfin, ceux à qui plaît le style « anti-establishment » du Président. Ils aiment son côté direct et sa critique du système washingtonien. Quels sont les avantages et inconvénients de Joe Biden ? Son avantage est qu’il est en avance dans des groupes électoraux très différents : les jeunes, fortement mobilisés, les seniors, habituellement républicains et qui semblent contrariés par la politique de Donald Trump, les femmes et les minorités. Et surtout, il a de nombreux arguments liés à la gestion de Trump sur la Covid-19. Son désavantage est que la plupart de ses électeurs votent essentiellement « contre Trump » et pas « pour Biden ». C’est un déficit d’enthousiasme indéniable. Contrairement à son prédécesseur, Obama, Trump n’est venu qu’une fois en Afrique ; ce qui est ressenti comme un manque d’intérêt pour le continent. Biden pourrait-il être un meilleur acteur dans les relations Afrique-Amérique ? Peut-on faire pire ? Trump n’a eu aucune politique africaine. Il n’a même pas cherché à montrer un quelconque intérêt. Joe Biden pourrait renouer avec un certain nombre de politiques initiées par Barack Obama, notamment en matière d’électrification du continent. Mais, on attend surtout de lui qu’il soutienne le multilatéralisme qui est la condition d’émergence d’une Union africaine solide. Comment le nationalisme et la loi du plus fort pourraient-ils être des atouts pour le continent africain ? Dans plusieurs villes africaines, Trump est vu comme quelqu’un d’hostile aux minorités (noire notamment) et Biden comme un progressiste. Est-ce si manichéen ? Non, ce n’est pas si manichéen, car Trump affirme que sa politique a été bénéfique à la communauté afro-américaine. Avant la Covid-19, le taux de chômage de la communauté était très bas. Mais, comment oublier ses propos suite aux manifestations de Charlottesville où il a placé sur le même plan les militants d’extrême droite et les antiracistes ? Comment oublier son absence d’empathie suite à la mort de George Floyd et sa manière de dénigrer Black lives matter, en assimilant les mouvements de colère à l’extrême gauche « Antifa ». Quel candidat pourrait être « la meilleure chose » aux nombreux Sénégalais sans-papiers aux États-Unis ? Sur la question de l’émigration, Joe Biden a une position plus ouverte que Donald Trump. L’aile gauche du parti démocrate, favorable à une politique de citoyenneté délivrée aux migrants, promet d’être en position forte dans la nouvelle équipe exécutive. Qui sera vainqueur de l’élection ? La position de Joe Biden est assurément plus forte que celle d’Hillary Clinton il y a quatre ans. Il n’est donc pas sûr que l’histoire se répète avec, une nouvelle fois, une surprise nommée Trump. Cependant, il faut rester prudent concernant les sondages. On parle particulièrement des éventuels « électeurs cachés » de Trump : des gens qui n’oseraient pas avouer leur vote aux sondeurs… Ce n’est pas moi qui le dit, mais la cheffe de campagne de Joe Biden : jusqu’au bout, il faudra rester prudent. Si Biden veut gagner l’élection dès le 3 novembre, sans qu’on doive se lancer dans un long et périlleux dépouillement (et recomptage…), il lui faudra remporter la Floride. Autre possibilité : qu’il l’emporte dans un des États promis aux républicains (Texas, Nevada ou Géorgie…). 
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           Le principal adversaire du chef de l’Etat tanzanien sortant John Magufuli à la présidentielle du 28 octobre, Tundu Lissu, a été arrêté lundi, après sept autres responsables de son parti d’opposition, le Chadema, a annoncé la police. « Nous l’avons arrêté à Dar es Salaam, alors qu’il sortait des bureaux de l’Union européenne », a indiqué à l’AFP le chef de la police de la capitale économique, Lazaro Mambosasa, « en lien avec les manifestations interdites », auxquelles des responsables du Chadema, dont M. Lissu, battu lors du scrutin, avaient appelé pour protester contre des élections marquées par la fraude, selon eux. AFP
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       Le gouvernement sénégalais « condamne fermement » les actes terroristes et toute autre forme de violence, y compris morale. C’est ce que révèle un communiqué du ministère des Affaires étrangères et des Sénégalais de l’extérieur, daté du 31 octobre 2020.\n« Le Sénégal a suivi avec préoccupation les attentats des 16, 29 et 31 octobre 2020, perpétrés respectivement à Conflans-Sainte-Honorine, à Nice et à Lyon, en France », précise le document.\nLe Sénégal rappelle, également, « son attachement aux valeurs de paix, de fraternité humaine et de respect du culte et du sentiment religieux, comme fondements essentiels de la coexistence pacifique des peuples et des civilisations ».
#>   press_article
#> 1      lesoleil
#> 2      lesoleil
#> 3      lesoleil
#> 4      lesoleil
#> 5      lesoleil
#> 6      lesoleil